@@ -72,7 +72,7 @@
     "is-ci":"2.0.0",
     "latest-version": "^4.0.0",
     "semver": "^5.5.1",
-    "tslib": "^1.9.3"
+    "tslib": "^2.0.3"
   },
   "devDependencies": {
     "@oclif/dev-cli": "^1.22.0",
@@ -89,8 +89,8 @@
     "nock": "^11.7.0",
     "nyc": "^14.1.1",
     "rimraf": "^2.6.3",
-    "ts-node": "^8.1.0",
+    "ts-node": "^9.0.0",
     "tslint": "^5.16.0",
-    "typescript": "^3.5.3"
+    "typescript": "^4.0.3"
   }
 }
