@@ -15,8 +15,8 @@
   "main": "lib/index.js",
   "repository": "https://github.com/microsoft/botframework-cli",
   "scripts": {
-    "prepack": "",
     "postpack": "",
+    "prepack": "npm run clean && npm run build",
     "posttest": "",
     "build": "tsc -b",
     "clean": "rimraf ./.nyc_output ./lib ./package-lock.json ./tsconfig.tsbuildinfo",
@@ -36,7 +36,7 @@
     "cli-ux": "~4.9.3",
     "debug": "^4.1.1",
     "fs-extra": "^7.0.1",
-    "tslib": "~1.10.0"
+    "tslib": "^2.0.3"
   },
   "devDependencies": {
     "@oclif/plugin-help": "^2.1.6",
@@ -60,8 +60,8 @@
     "rimraf": "^2.6.3",
     "source-map-support": "~0.5.16",
     "testdouble": "^3.11.0",
-    "ts-node": "^8.1.0",
+    "ts-node": "^9.0.0",
     "tslint": "^5.16.0",
-    "typescript": "^3.5.3"
+    "typescript": "^4.0.3"
   }
 }
