@@ -26,6 +26,7 @@
   "repository": "https://github.com/microsoft/botframework-cli/tree/master/packages/config",
   "scripts": {
     "postpack": "rimraf oclif.manifest.json",
+    "prepack": "npm run clean && npm run build && npm run doc:readme",
     "posttest": "tslint -p test -t stylish",
     "build": "tsc -b",
     "clean": "rimraf ./.nyc_output ./lib ./package-lock.json ./tsconfig.tsbuildinfo",
@@ -43,7 +44,7 @@
     "@oclif/errors": "~1.2.2",
     "cli-ux": "^5.3.0",
     "fs-extra": "^7.0.1",
-    "tslib": "^1.10.0"
+    "tslib": "^2.0.3"
   },
   "devDependencies": {
     "@oclif/dev-cli": "^1.22.0",
@@ -58,8 +59,8 @@
     "mocha": "^6.2.2",
     "nyc": "^14.1.1",
     "rimraf": "^2.6.3",
-    "ts-node": "^8.3.0",
+    "ts-node": "^9.0.0",
     "tslint": "^5.18.0",
-    "typescript": "^3.5.3"
+    "typescript": "^4.0.3"
   }
 }
