@@ -27,6 +27,7 @@
   "repository": "https://github.com/microsoft/botframework-cli/tree/master/packages/dialog",
   "scripts": {
     "postpack": "rimraf oclif.manifest.json",
+    "prepack": "npm run clean && npm run build && npm run doc:readme",
     "posttest": "tslint -p test -t stylish",
     "build": "tsc -b",
     "clean": "rimraf ./.nyc_output ./lib ./package-lock.json ./tsconfig.tsbuildinfo",
@@ -53,7 +54,7 @@
     "os": "~0.1.1",
     "path": "^0.12.7",
     "seedrandom": "~3.0.5",
-    "tslib": "^1.10.0",
+    "tslib": "^2.0.3",
     "xml2js": "^0.4.19",
     "json-ptr": "~1.3.0",
     "json-schema-merge-allof": "~0.7.0",
@@ -73,8 +74,8 @@
     "mocha": "^6.2.2",
     "nyc": "^14.1.1",
     "rimraf": "^2.6.3",
-    "ts-node": "^8.3.0",
+    "ts-node": "^9.0.0",
     "tslint": "^5.18.0",
-    "typescript": "^3.5.3"
+    "typescript": "^4.0.3"
   }
 }
