@@ -16,7 +16,7 @@
   "license": "MIT",
   "repository": "https://github.com/microsoft/botframework-cli/tree/master/packages/lu",
   "scripts": {
-    "postpack": "",
+    "prepack": "npm run clean && npm run build",
     "posttest": "tslint -p test -t stylish",
     "build": "tsc -b",
     "clean": "rimraf ./.nyc_output ./lib ./package-lock.json ./tsconfig.tsbuildinfo",
@@ -35,8 +35,6 @@
   "dependencies": {
     "@azure/cognitiveservices-luis-authoring": "4.0.0-preview.1",
     "@azure/ms-rest-azure-js": "2.0.1",
-    "@oclif/command": "~1.5.19",
-    "@oclif/errors": "~1.2.2",
     "@types/node-fetch": "~2.5.5",
     "antlr4": "^4.7.2",
     "chalk": "2.4.1",
@@ -50,11 +48,9 @@
     "lodash": "^4.17.19",
     "node-fetch": "~2.6.0",
     "semver": "^5.5.1",
-    "tslib": "^1.10.0"
+    "tslib": "^2.0.3"
   },
   "devDependencies": {
-    "@oclif/test": "^1.2.5",
-    "@oclif/tslint": "^3.1.1",
     "@types/chai": "^4.2.0",
     "@types/lodash": "~4.14.159",
     "@types/mocha": "^5.2.7",
@@ -65,9 +61,9 @@
     "nock": "^11.7.0",
     "nyc": "^14.1.1",
     "rimraf": "^2.6.3",
-    "ts-node": "^8.3.0",
+    "ts-node": "^9.0.0",
     "tslint": "^5.18.0",
-    "typescript": "^3.5.3",
+    "typescript": "^4.0.3",
     "uuid": "^3.3.3"
   }
 }
