@@ -9,7 +9,7 @@
     "@oclif/config": "~1.13.3",
     "@oclif/plugin-plugins": "~1.7.9",
     "cli-ux": "^5.4.1",
-    "tslib": "^1.10.0"
+    "tslib": "^2.0.3"
   },
   "devDependencies": {
     "@oclif/dev-cli": "^1.22.2",
@@ -25,8 +25,8 @@
     "globby": "^10.0.1",
     "mocha": "^5.2.0",
     "nyc": "^14.1.1",
-    "ts-node": "^8.5.4",
-    "typescript": "^3.5.3",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.3",
     "rimraf": "^2.6.3"
   },
   "engines": {
@@ -53,7 +53,7 @@
   "scripts": {
     "postpack": "rimraf oclif.manifest.json",
     "posttest": "eslint . --ext .ts --config .eslintrc",
-    "prepack": "rimraf lib && tsc -b && oclif-dev manifest && oclif-dev readme",
+    "prepack": "npm run clean && npm run build && oclif-dev manifest && oclif-dev readme",
     "build": "tsc -b",
     "clean": "rimraf ./.nyc_output ./lib ./package-lock.json ./tsconfig.tsbuildinfo",
     "test": "mocha",
