@@ -5,12 +5,12 @@
 // tslint:disable:no-console
 // tslint:disable:no-object-literal-type-assertion
 
-import {assert} from 'chai'
+import { assert } from 'chai'
 import * as fs from 'fs-extra'
 import 'mocha'
 import * as os from 'os'
 import * as ppath from 'path'
-import SchemaMerger from '../../../src/library/schemaMerger'
+import * as merger from '../../../src/library/schemaMerger'
 let srcDir = ppath.resolve('test/commands/dialog/')
 let tempDir = ppath.join(os.tmpdir(), 'test.out')
 
@@ -24,21 +24,24 @@ function countMatches(pattern: string | RegExp, lines: string[]): number {
     return count
 }
 
-async function merge(patterns: string[], output?: string, verbose?: boolean, schemaPath?: string): Promise<[boolean, string[]]> {
+async function merge(patterns: string[], output?: string, verbose?: boolean, schemaPath?: string, checkOnly?: boolean): Promise<[merger.Imports | undefined, string[]]> {
     let lines: string[] = []
-    let logger = msg => {
+    let logger = (msg: string) => {
         console.log(msg)
         lines.push(msg)
     }
-    let merger = new SchemaMerger(patterns,
-        output ? ppath.join(tempDir, output) : '',
+    let outputDir = output ? ppath.join(tempDir, output) : ''
+    let mergeClass = new merger.SchemaMerger(patterns,
+        outputDir,
+        undefined,
+        checkOnly === undefined ? false : checkOnly,
         verbose || false,
         logger, logger, logger,
-        undefined, 
-        schemaPath ? ppath.join(tempDir, schemaPath) : undefined, 
+        undefined,
+        schemaPath ? ppath.join(tempDir, schemaPath) : undefined,
         false,
         ppath.join(srcDir, 'nuget'))
-    let merged = await merger.merge()
+    let merged = await mergeClass.merge()
     return [merged, lines]
 }
 
@@ -79,8 +82,52 @@ async function compareToOracle(name: string, oraclePath?: string): Promise<objec
     return generated
 }
 
+async function modifyFile(path: string, pattern: RegExp, replacement: string) {
+    let contents = await fs.readFile(path, 'utf-8')
+    contents = contents.replace(pattern, replacement)
+    await fs.writeFile(path, contents)
+}
+
+function checkMerged(merged: merger.Imports | undefined, adds: number, conflicts: number, deletes: number, unchanged: number, msg: string, components?: any[]) {
+    if (merged) {
+        // NOTE: This is useful to help write tests.
+        // console.log(`adds: ${merged.added.length}, conflicts: ${merged.conflicts.length}, deleted: ${merged.deleted.length}, unchanged: ${merged.unchanged.length}`)
+        // console.log(JSON.stringify(merged.components))
+        assert(merged.added.length === adds, `Wrong number of adds ${msg}`)
+        assert(merged.conflicts.length === conflicts, `Wrong number of conflicts ${msg}`)
+        assert(merged.deleted.length === deletes, `Wrong number of deletes ${msg}`)
+        assert(merged.unchanged.length === unchanged, `Wrong number of unchanged ${msg}`)
+        if (components) {
+            assert(merged.components.length === components.length, `Wrong number of components ${msg}`)
+            for (let i = 0; i < components.length; ++i) {
+                let actual: any = merged.components[i]
+                let expected: any = components[i]
+                for (let key of Object.keys(expected)) {
+                    let actualVal = actual[key]
+                    let expectedVal = expected[key]
+                    if (Array.isArray(expectedVal)) {
+                        assert(actualVal.length === expectedVal.length, `${actual.name}.${key} length ${actualVal.length} != ${expectedVal.length} ${msg}`)
+                        for (let e = 0; e < actualVal.length; ++e) {
+                            assert(actualVal[e] === expectedVal[e], `${actual.name}.${key}[${e}] ${actualVal[e]} != ${expectedVal[e]} ${msg}`)
+                        }
+                    } else {
+                        if (key === 'path') {
+                            actualVal = actualVal.replace(/\\/g, '/')
+                            expectedVal = expectedVal.replace(/\\/g, '/')
+                            assert(actualVal.endsWith(expectedVal),
+                                `${actual.name}.${key} ${actualVal} does not end with ${expectedVal} ${msg}`)
+                        } else {
+                            assert(actualVal === expectedVal, `${actual.name}.${key} ${actualVal} != ${expectedVal} ${msg}`)
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
 describe('dialog:merge', async () => {
-    before(async () => {
+    beforeEach(async () => {
         // If you want to regenerate the oracle *.schema files, run schemas/makeschemas.cmd
         await fs.remove(tempDir)
         await fs.mkdirp(tempDir)
@@ -165,16 +212,27 @@ describe('dialog:merge', async () => {
         let [merged, lines] = await merge(['projects/project3/project3.csproj'], 'project3.schema', true)
         assert(countMatches(/error|warning/i, lines) === 0, 'Should not have got errors')
         assert(merged, 'Could not merge')
-        assert(countMatches(/Following.*project3/, lines) === 1, 'Did not follow project1')
+        assert(countMatches(/Following.*project3/, lines) === 1, 'Did not follow project3')
         assert(countMatches(/Following nuget.*nuget3.*1.0.0/, lines) === 1, 'Did not follow nuget3')
         assert(countMatches(/Parsing.*nuget3.component1.schema/, lines) === 1, 'Missing nuget3.component1.schema')
-        assert(countMatches(/Copying/i, lines) === 4, 'Wrong number of copies')
-        assert(countMatches(/Copying.*nuget3.lg/i, lines) === 1, 'Did not copy .lg')
-        assert(countMatches(/Copying.*nuget3.lu/i, lines) === 1, 'Did not copy .lu')
-        assert(countMatches(/Copying.*nuget3.qna/i, lines) === 1, 'Did not copy .qna')
-        assert(await fs.pathExists(ppath.join(tempDir, 'generated', 'nuget3', 'assets', 'nuget3.qna')), 'Did not copy directory')
+        assert(countMatches(/Copying/i, lines) === 2, 'Wrong number of copies')
+        assert(countMatches(/Copy /i, lines) === 6, 'Did not copy resources')
+        assert(await fs.pathExists(ppath.join(tempDir, 'imported', 'nuget3', 'stuff', 'nuget3.qna')), 'Did not copy directory')
         await compareToOracle('project3.schema')
         await compareToOracle('project3.en-us.uischema')
+        checkMerged(merged, 6, 0, 0, 0, '',
+            [{
+                name: 'nuget3',
+                version: '1.0.0',
+                path: 'nuget\\nuget3\\1.0.0\\nuget3.nuspec',
+                description: 'Nuget 3',
+                releaseNotes: 'Changed metatdata',
+                authors: ['Chris Tom', 'John Mark'],
+                keywords: ['a', 'good', 'thing'],
+                icon: 'icon.png', repository: 'https://github.com',
+                license: 'MIT',
+                language: 'en-us', copyright: 'Mine only.', includesSchema: true, includesExports: true
+            }])
     })
 
     it('csproj-errors', async () => {
@@ -215,6 +273,93 @@ describe('dialog:merge', async () => {
         await compareToOracle('project5.en-us.uischema')
     })
 
+    it('csproj-schema', async () => {
+        console.log('\nStart csproj-schema')
+        let [merged, lines] = await merge(['projects/project3/project3.csproj'], 'project3.schema', false)
+        assert(countMatches(/error|warning/i, lines) === 0, 'Should not have got errors')
+        assert(merged, 'Could not merge')
+        let [merged2, lines2] = await merge(['projects/project3/project3.csproj'], 'project3-schema.schema', true, 'project3.schema')
+        assert(countMatches(/error|warning/i, lines2) === 0, 'Should not have got errors')
+        assert(countMatches(/using merged schema/i, lines2) === 1, 'Should use merged schema')
+        assert(merged2, 'Could not merge')
+        await compareToOracle('project3-schema.en-us.uischema', 'project3.en-us.uischema')
+    })
+
+    it('csproj-import', async () => {
+        console.log('\nStart csproj-import')
+        let project = ppath.join(tempDir, 'project3.csproj')
+        await fs.emptyDir(tempDir)
+        await fs.copyFile('projects/project3/project3.csproj', project)
+
+        // First import
+        console.log('\nFirst import')
+        let [merged, lines] = await merge([project], 'project3.schema', false)
+        assert(merged, 'Could not merge')
+        assert(countMatches(/error|warning/i, lines) === 0, 'Error merging schemas')
+        checkMerged(merged, 6, 0, 0, 0, 'initial')
+
+        // Second import with no changes
+        console.log('\nSecond import without changes')
+        let [merged2, lines2] = await merge([project], 'project3.schema', false)
+        assert(merged2, 'Could not merge 2nd')
+        assert(countMatches(/error|warning/i, lines2) === 0, 'Error merging schemas 2nd')
+        checkMerged(merged2, 1, 0, 0, 5, '2nd')
+
+        // Third import with changes but check only
+        console.log('\nThird import with check-only changes')
+        let luPath = ppath.join(tempDir, 'imported/nuget3/nuget3.lu')
+        let jpgPath = ppath.join(tempDir, 'imported/nuget3/nuget3.jpg')
+        let deletedPath = ppath.join(tempDir, 'imported/nuget3/stuff/nuget3-deleted.dialog')
+        await modifyFile(project, /1.0.0/, '1.0.1')
+        await modifyFile(luPath, /intent/, 'intent modified')
+        await modifyFile(ppath.join(tempDir, 'imported/nuget3/nuget3.lg'), /template/, 'template modified')
+        await modifyFile(ppath.join(tempDir, 'imported/nuget3/stuff/nuget3.qna'), /question/, 'question modified')
+        await modifyFile(ppath.join(tempDir, 'imported/nuget3/stuff/nuget3.dialog'), /dialog/, 'dialog modified')
+        await modifyFile(jpgPath, /picture/, 'picture modified')
+        let [merged3, lines3] = await merge([project], 'project3.schema', true, undefined, true)
+        assert(merged3, 'Could not merge 3rd')
+        assert(countMatches(/error/i, lines3) === 0, 'Error merging schemas 3rd')
+        assert(countMatches(/warning/i, lines3) === 3, 'Wrong number of warnings 3rd')
+        checkMerged(merged3, 1, 3, 1, 1, '3rd')
+        assert(countMatches('modified', lines3) === 0, 'Missed deletion change 3rd')
+        assert(countMatches('conflicting', lines3) === 3, 'Missed conflicts 3rd')
+        assert((await fs.readFile(luPath, 'utf8')).includes('modified'), 'Wrote file in check-only')
+        assert((await fs.readFile(jpgPath, 'utf8')).includes('modified'), 'Wrote file in check-only')
+        assert(await fs.pathExists(deletedPath), 'Deleted file in check-only')
+
+        // Fourth import with changes
+        console.log('\nFourth import with changes')
+        await modifyFile(deletedPath, /dialog/, 'dialog modified')
+        let [merged4, lines4] = await merge([project], 'project3.schema', true, undefined, false)
+        assert(merged4, 'Could not merge 4th')
+        assert(countMatches(/error/i, lines4) === 0, 'Error merging schemas 4th')
+        assert(countMatches(/warning/i, lines4) === 4, 'Wrong number of warnings 4th')
+        assert(merged4?.added.length === 1, 'Wrong number added 4th')
+        assert(merged4?.deleted.length === 0, 'Wrong number deleted 4th')
+        assert(merged4?.unchanged.length === 1, 'Wrong number unchanged 4th')
+        assert(merged4?.conflicts.length === 4, 'Wrong number of conflicts on 4th')
+        assert(countMatches('modified', lines4) === 1, 'Missed deletion change 4th')
+        assert(countMatches('conflicting', lines4) === 3, 'Missed conflicts 4th')
+        assert(countMatches('Unchanged', lines4) === 1, 'Missed unchanged 4th')
+        assert(countMatches('deleted', lines4) === 1, 'Missed delete 4th')
+        assert((await fs.readFile(luPath, 'utf8')).includes('changed'), 'Did not write file')
+        assert((await fs.readFile(jpgPath, 'utf8')).includes('changed'), 'Wrote file in check-only')
+        assert(!await fs.pathExists(deletedPath), 'Did not delete file')
+
+        // Fifth import with component removed
+        console.log('\nFifth import removing reference')
+        await modifyFile(project, /<PackageReference.*/, '')
+        let [merged5, lines5] = await merge([project], 'project3.schema', true, undefined, false)
+        assert(merged5, 'Could not merge 5th')
+        assert(countMatches(/error/i, lines5) === 1, 'Error merging schemas 5th')
+        assert(countMatches(/warning/i, lines5) === 1, 'Wrong number of warnings 5th')
+        assert(merged5?.added.length === 0, 'Wrong number added 5th')
+        assert(merged5?.deleted.length === 4, 'Wrong number deleted 5th')
+        assert(merged5?.unchanged.length === 0, 'Wrong number unchanged 5th')
+        assert(merged5?.conflicts.length === 1, 'Wrong number of conflicts on 5th')
+        assert(!await fs.pathExists(ppath.join(tempDir, 'imported/nuget3')))
+    })
+
     it('package.json', async () => {
         console.log('\nStart package.json')
         let [merged, lines] = await merge(['npm/node_modules/root-package/package.json'], 'root-package.schema', true)
@@ -224,8 +369,44 @@ describe('dialog:merge', async () => {
         assert(countMatches('dependent-package.schema', lines) === 1, 'Missing dependent-package.schema')
         assert(countMatches('parent-package.schema', lines) === 1, 'Missing parent-package.schema')
         assert(countMatches('no-package.schema', lines) === 0, 'Extra no-package.schema')
+        assert(countMatches('Copy ', lines) === 5, 'Wrong number of copies')
+        assert(await fs.pathExists(ppath.join(tempDir, 'imported', 'dependent-package', 'assets', 'dependent-package.jpg')), 'Incomplete assets copy')
+        assert(!await fs.pathExists(ppath.join(tempDir, 'imported', 'root-package')), 'Copied rootx')
         await compareToOracle('root-package.schema')
         await compareToOracle('root-package.uischema')
+        checkMerged(merged, 5, 0, 0, 0, '',
+            [{
+                name: 'dependent-package',
+                version: '1.0.0',
+                path: 'npm\\node_modules\\root-package\\node_modules\\dependent-package\\package.json',
+                description: 'A package',
+                releaseNotes: 'A good time.',
+                authors: ['Chris Tom'],
+                keywords: ['somthing wicked', 'this way'],
+                icon: 'icon.png',
+                repository: 'https://github.com',
+                license: 'MIT',
+                language: 'en-us',
+                copyright: 'mine',
+                includesSchema: true,
+                includesExports: true
+            },
+            {
+                name: 'parent-package',
+                version: '1.0.0',
+                path: 'npm\\node_modules\\parent-package\\package.json',
+                description: '',
+                releaseNotes: '',
+                authors: ['Chris Tom, tom@botmail.com, https://christom.com'],
+                keywords: [],
+                icon: '',
+                repository: 'https://github.com',
+                license: '',
+                language: '',
+                copyright: '',
+                includesSchema: true,
+                includesExports: false
+            }])
     })
 
     it('nuspec', async () => {
@@ -252,17 +433,15 @@ describe('dialog:merge', async () => {
         }
     })
 
-    it('csproj -schema', async () => {
-        console.log('\nStart csproj')
-        let [merged, lines] = await merge(['projects/project3/project3.csproj'], 'project3.schema', false)
-        assert(countMatches(/error|warning/i, lines) === 0, 'Should not have got errors')
-        assert(merged, 'Could not merge')
-        let [merged2, lines2] = await merge(['projects/project3/project3.csproj'], 'project3-schema.schema', true, 'project3.schema')
-        assert(countMatches(/error|warning/i, lines2) === 0, 'Should not have got errors')
-        assert(countMatches(/using merged schema/i, lines2) === 1, 'Should use merged schema')
-        assert(merged2, 'Could not merge')
-        await compareToOracle('project3-schema.en-us.uischema', 'project3.en-us.uischema')
+    /* Example of invoking library through test because of issues with launch.json
+    it('botbuilder-schema', async () => {
+        console.log('\nStart botbuilder-schema')
+        let [merged, lines] = await merge(['../../../../../../botbuilder-dotnet/libraries/**\*.schema',
+    '../../../../../../botbuilder-dotnet/tests/**\*.schema'], 'sdk.schema', true)
+        console.log(merged)
+        console.log(lines)
     })
+    */
 })
 
 /* TODO: These tests are related to verify and need to be updated and moved there.
