@@ -3,31 +3,17 @@
  * Licensed under the MIT License.
  */
 
-const fs = require('fs-extra')
 const path = require('path')
 const file = require('../../utils/filehelper')
 const fileExtEnum = require('../utils/helpers').FileExtTypeEnum
 const crossTrainer = require('./crossTrainer')
-const confighelper = require('./confighelper')
 
 module.exports = {
-  /**
-   * Generate cross train config based on input folder and root dialog file.
-   * @param {string} inputFolder full path of input lu and qna files folder.
-   * @param {string} rootDialogFile full path of root dialog file.
-   * @returns {string} config object json string.
-   */
-  generateConfig: async function (inputFolder, rootDialogFile) {
-    const configStr = await confighelper.generateConfig(inputFolder, rootDialogFile)
-
-    return configStr
-  },
-
   /**
    * Cross train lu and qna files.
    * @param {string} input full path of input lu and qna files folder.
    * @param {string} intentName interruption intent name. Default value is _Interruption.
-   * @param {string} config path to config of mapping rules or mapping rules json content itself. If undefined, it will read config.json from input folder.
+   * @param {string} config path to config file of mapping rules.
    * @param {boolean} verbose verbose to indicate whether log warnings and errors or not when parsing cross-train files.
    * @returns {luResult: any, qnaResult: any} trainedResult of luResult and qnaResult or undefined if no results.
    */
@@ -35,12 +21,41 @@ module.exports = {
     // Get all related file content.
     const luContents = await file.getFilesContent(input, fileExtEnum.LUFile)
     const qnaContents = await file.getFilesContent(input, fileExtEnum.QnAFile)
-    const configContent = config && !fs.existsSync(config) ? {id: path.join(input, 'config.json'), content: config} : await file.getConfigContent(config)
+    const configContent = await file.getConfigContent(config)
 
-    const configObject = file.getConfigObject(configContent, intentName, verbose)
+    let importResolver = async function (_, idsToFind) {
+      let importedContents = []
+      for (let idx = 0; idx < idsToFind.length; idx++) {
+        let file = idsToFind[idx]
+        if (path.isAbsolute(file.filePath)) {
+          if (file.filePath.endsWith(fileExtEnum.LUFile)) {
+            importedContents.push(...await file.getFilesContent(file.filePath, fileExtEnum.LUFile))
+          } else if (file.filePath.endsWith(fileExtEnum.QnAFile)) {
+            importedContents.push(...await file.getFilesContent(file.filePath, fileExtEnum.QnAFile))
+          }
+        } else {
+          const fileName = path.basename(file.filePath)
+          if (fileName.endsWith(fileExtEnum.LUFile)) {
+            importedContents.push(...luContents.filter(luContent => luContent.id === path.basename(fileName, fileExtEnum.LUFile)))
+          } else if (fileName.endsWith(fileExtEnum.QnAFile)) {
+            importedContents.push(...qnaContents.filter(qnaContent => qnaContent.id === path.basename(fileName, fileExtEnum.QnAFile)))
+          }
+        }
+      }
 
-    const trainedResult = await crossTrainer.crossTrain(luContents, qnaContents, configObject)
-    
+      return importedContents
+    }
+
+    const trainedResult = await crossTrainer.crossTrain(
+      luContents,
+      qnaContents,
+      JSON.parse(configContent.content), {
+      configId: configContent.id,
+      intentName,
+      verbose,
+      importResolver
+    })
+
     return trainedResult
   }
 }
