@@ -3,42 +3,26 @@
  * Licensed under the MIT License.
  */
 
-import {Settings} from './settings'
-import * as path from 'path'
-
 export class Recognizer {
-  static load(luFile: string, targetFileName: string, dialogPath: string, luisSettings: Settings, existingRecognizer: any, schema?: string): Recognizer {
-    if (existingRecognizer) {
-      let recognizer = new Recognizer(luFile, targetFileName, schema)
-      recognizer.dialogPath = dialogPath
-      Object.assign(recognizer, existingRecognizer)
-      recognizer.setAppId(luisSettings.luis[path.basename(luFile).split('.').join('_')])
-
-      return recognizer
-    }
-
-    let recognizer = new Recognizer(luFile, targetFileName, schema)
-    recognizer.dialogPath = dialogPath
-
-    return recognizer
-  }
-
   versionId: string
   private readonly id: string
   private readonly applicationId: string | undefined
+  private readonly version: string | undefined
   private readonly endpoint: string | undefined
   private readonly endpointKey: string | undefined
   private readonly $schema: string | undefined
   private appId: string
   private dialogPath: string | undefined
 
-  constructor(private readonly luFile: string, targetFileName: string, schema?: string) {
+  constructor(private readonly luFile: string, targetFileName: string, dialogPath: string, schema?: string) {
     this.appId = ''
     this.id = `LUIS_${targetFileName.split('.')[0]}`
-    this.applicationId = `=settings.luis.${targetFileName.split('.').join('_').replace(/-/g, '_')}`
+    this.applicationId = `=settings.luis.${targetFileName.split('.').join('_').replace(/-/g, '_')}.appId`
     this.endpoint = '=settings.luis.endpoint'
     this.endpointKey = '=settings.luis.endpointKey'
-    this.versionId = '0.1'
+    this.version = `=settings.luis.${targetFileName.split('.').join('_').replace(/-/g, '_')}.version`
+    this.versionId = "0.1"
+    this.dialogPath = dialogPath
     this.$schema = schema
   }
 
@@ -47,6 +31,7 @@ export class Recognizer {
       $kind: 'Microsoft.LuisRecognizer',
       id: this.id,
       applicationId: this.applicationId,
+      version: this.version,
       endpoint: this.endpoint,
       endpointKey: this.endpointKey
     }
