@@ -8,6 +8,7 @@ const SimpleIntentSection = require('./simpleIntentSection');
 const EntitySection = require('./entitySection');
 const NewEntitySection =  require('./newEntitySection');
 const ImportSection = require('./importSection');
+const ReferenceSection = require('./referenceSection');
 const QnaSection = require('./qnaSection');
 const ModelInfoSection = require('./modelInfoSection');
 const LUErrorListener = require('./luErrorListener');
@@ -18,47 +19,64 @@ const Range = require('./diagnostic').Range;
 const Position = require('./diagnostic').Position;
 const NEWLINE = require('os').EOL;
 
+const defaultConfig = {
+  enableModelDescription: true,
+  enableComments: true // Temporarily enabled by default, cannot be configured
+}
+
 class LUParser {
 
     /**
-     * 
-     * @param {string} text 
-     * @param {LUResource} luResource 
+     *
+     * @param {string} text
+     * @param {LUResource} luResource
      */
-    static parseWithRef(text, luResource) {
+    static parseWithRef(text, luResource, config) {
+        config = config || {};
+        config = {...defaultConfig, ...config};
         if (text === undefined || text === '') {
             return new LUResource([], '', []);
         }
 
         const sectionEnabled = luResource ? this.isSectionEnabled(luResource.Sections) : undefined;
 
-        return this.parse(text, sectionEnabled);
+        return this.parse(text, sectionEnabled, config);
     }
 
     /**
      * @param {string} text
      */
-    static parse(text, sectionEnabled) {
+    static parse(text, sectionEnabled, config) {
+        config = config || {};
+        config = {...defaultConfig, ...config};
         if (text === undefined || text === '') {
             return new LUResource([], '', []);
         }
 
         let {fileContent, errors} = this.getFileContent(text);
 
-        return this.extractFileContent(fileContent, text, errors, sectionEnabled);
+        return this.extractFileContent(fileContent, text, errors, sectionEnabled, config);
     }
 
-    static extractFileContent(fileContent, content, errors, sectionEnabled) {
+    static extractFileContent(fileContent, content, errors, sectionEnabled, config) {
         let sections = [];
+        let modelInfoSections = [];
+
         try {
-            let modelInfoSections = this.extractModelInfoSections(fileContent);
-            modelInfoSections.forEach(section => errors = errors.concat(section.Errors));
-            sections = sections.concat(modelInfoSections);
+            modelInfoSections = this.extractModelInfoSections(fileContent);
         } catch (err) {
             errors.push(BuildDiagnostic({
                 message: `Error happened when parsing model information: ${err.message}`
             }))
         }
+
+        if (modelInfoSections && modelInfoSections.length > 0 && !config.enableModelDescription) {
+          errors.push(BuildDiagnostic({
+            message: `Do not support Model Description. Please make sure enableModelDescription is set to true.`
+          }))
+        }
+        modelInfoSections.forEach(section => errors = errors.concat(section.Errors));
+        sections = sections.concat(modelInfoSections);
 
         try {
             let isSectionEnabled = sectionEnabled === undefined ?  this.isSectionEnabled(sections) : sectionEnabled;
@@ -72,12 +90,12 @@ class LUParser {
                     let emptyIntentSection = new SimpleIntentSection();
                     emptyIntentSection.Name = section.Name;
                     emptyIntentSection.Id = `${emptyIntentSection.SectionType}_${emptyIntentSection.Name}`
-                    
+
                     // get the end character index
                     // this is default value
                     // it will be reset in function extractSectionBody()
                     let endCharacter = section.Name.length + 2;
-                   
+
                     const range = new Range(section.Range.Start, new Position(section.Range.Start.Line, endCharacter))
                     emptyIntentSection.Range = range;
                     let errorMsg = `no utterances found for intent definition: "# ${emptyIntentSection.Name}"`
@@ -143,6 +161,16 @@ class LUParser {
         }
 
         try {
+            let referenceSections = this.extractReferenceSections(fileContent);
+            referenceSections.forEach(section => errors = errors.concat(section.Errors));
+            sections = sections.concat(referenceSections);
+        } catch (err) {
+            errors.push(BuildDiagnostic({
+                message: `Error happened when parsing reference section: ${err.message}`
+            }))
+        }
+
+        try {
             let qnaSections = this.extractQnaSections(fileContent);
             qnaSections.forEach(section => errors = errors.concat(section.Errors));
             sections = sections.concat(qnaSections);
@@ -166,7 +194,7 @@ class LUParser {
         if (text === undefined
             || text === ''
             || text === null) {
-            
+
             return undefined;
         }
 
@@ -180,13 +208,13 @@ class LUParser {
         parser.addErrorListener(listener);
         parser.buildParseTrees = true;
         const fileContent = parser.file();
-        
+
         return { fileContent, errors };
     }
 
     /**
      * @param {FileContext} fileContext
-     * @param {string} content 
+     * @param {string} content
      */
     static extractNestedIntentSections(fileContext, content) {
         if (fileContext === undefined
@@ -204,8 +232,8 @@ class LUParser {
     }
 
     /**
-     * @param {FileContext} fileContext 
-     * @param {string} content 
+     * @param {FileContext} fileContext
+     * @param {string} content
      */
     static extractSimpleIntentSections(fileContext, content) {
         if (fileContext === undefined
@@ -223,7 +251,7 @@ class LUParser {
     }
 
     /**
-     * @param {FileContext} fileContext 
+     * @param {FileContext} fileContext
      */
     static extractEntitiesSections(fileContext) {
         if (fileContext === undefined
@@ -241,7 +269,7 @@ class LUParser {
     }
 
     /**
-     * @param {FileContext} fileContext 
+     * @param {FileContext} fileContext
      */
     static extractNewEntitiesSections(fileContext) {
         if (fileContext === undefined
@@ -252,14 +280,14 @@ class LUParser {
         let newEntitySections = fileContext.paragraph()
             .map(x => x.newEntitySection())
             .filter(x => x && x.newEntityDefinition());
-        
+
         let newEntitySectionList = newEntitySections.map(x => new NewEntitySection(x));
 
         return newEntitySectionList;
     }
 
     /**
-     * @param {FileContext} fileContext 
+     * @param {FileContext} fileContext
      */
     static extractImportSections(fileContext) {
         if (fileContext === undefined
@@ -277,7 +305,25 @@ class LUParser {
     }
 
     /**
-     * @param {FileContext} fileContext 
+     * @param {FileContext} fileContext
+     */
+    static extractReferenceSections(fileContext) {
+        if (fileContext === undefined
+            || fileContext === null) {
+                return [];
+        }
+
+        let referenceSections = fileContext.paragraph()
+            .map(x => x.referenceSection())
+            .filter(x => x !== undefined && x !== null);
+
+        let referenceSectionList = referenceSections.map(x => new ReferenceSection(x));
+
+        return referenceSectionList;
+    }
+
+    /**
+     * @param {FileContext} fileContext
      */
     static extractQnaSections(fileContext) {
         if (fileContext === undefined
@@ -295,7 +341,7 @@ class LUParser {
     }
 
     /**
-     * @param {FileContext} fileContext 
+     * @param {FileContext} fileContext
      */
     static extractModelInfoSections(fileContext) {
         if (fileContext === undefined
@@ -313,7 +359,7 @@ class LUParser {
     }
 
     /**
-     * @param {any[]} sections 
+     * @param {any[]} sections
      */
     static reconstractIntentSections(sections) {
         let newSections = []
@@ -334,7 +380,7 @@ class LUParser {
                     simpleIntentSections[simpleIntentSections.length - 1].Errors.push(...sections[index + 1].Errors)
                     index++
 
-                    while (index + 1 < sections.length 
+                    while (index + 1 < sections.length
                         && (sections[index + 1].SectionType === SectionType.ENTITYSECTION
                         || sections[index + 1].SectionType === SectionType.NEWENTITYSECTION
                         || (sections[index + 1].SectionType === SectionType.SIMPLEINTENTSECTION && sections[index + 1].IntentNameLine.includes('##')))) {
