@@ -66,13 +66,13 @@ const parseIntentsToLu = function(luisObj, luisJSON){
         if (luisJSON.test === true) {
             fileContent += `> Utterance passed in this intent: ${intent.intent.passNumber}/${intent.intent.count}` + NEWLINE
         }
-        fileContent += '## ' + intent.intent.name + NEWLINE;
+        fileContent += '# ' + intent.intent.name + NEWLINE;
         fileContent += parseUtterancesToLu(intent.utterances, luisJSON)
         fileContent += NEWLINE + NEWLINE;
         if (intent.intent.features) {
             let rolesAndFeatures = addRolesAndFeatures(intent.intent);
             if (rolesAndFeatures !== '') {
-                fileContent += `@ intent ${intent.intent.name}`;
+                fileContent += `@ intent ${intent.intent.name.includes(' ') ? `"${intent.intent.name}"` : `${intent.intent.name}`}`;
                 fileContent += rolesAndFeatures;
                 fileContent += NEWLINE + NEWLINE;
             }
