@@ -3,15 +3,20 @@
  * Licensed under the MIT License.
  */
 
-const retCode = require('./../utils/enums/CLI-errors')
-const exception = require('./../utils/exception')
 const {ServiceBase} = require('./serviceBase')
 const NEWLINE = require('os').EOL
+
+const absoluteUrlPattern = /^https?:\/\//i
 
 export class QnaBuildCore {
   private readonly service: any
 
   constructor(subscriptionkey: string, endpoint: string) {
+    // check endpoint is absolute or not
+    if (!absoluteUrlPattern.test(endpoint)) {
+      throw new Error(`Only absolute URLs are supported. "${endpoint}" is not an absolute qnamaker endpoint URL.`)
+    }
+
     this.service = new ServiceBase(endpoint, subscriptionkey)
   }
 
@@ -20,7 +25,7 @@ export class QnaBuildCore {
     const text = await response.text()
     const kbList = JSON.parse(text)
     if (kbList.error) {
-      throw (new exception(retCode.errorCode.LUIS_API_CALL_FAILED, text))
+      throw new Error(kbList.error.message)
     }
 
     return kbList
@@ -31,7 +36,7 @@ export class QnaBuildCore {
     const text = await response.text()
     const kb = JSON.parse(text)
     if (kb.error) {
-      throw (new exception(retCode.errorCode.LUIS_API_CALL_FAILED, text))
+      throw new Error(kb.error.message)
     }
 
     return kb
@@ -42,7 +47,7 @@ export class QnaBuildCore {
     const text = await response.text()
     const status = JSON.parse(text)
     if (status.error) {
-      throw (new exception(retCode.errorCode.LUIS_API_CALL_FAILED, text))
+      throw new Error(status.error.message)
     }
 
     return status
@@ -53,7 +58,7 @@ export class QnaBuildCore {
     const text = await response.text()
     const status = JSON.parse(text)
     if (status.error) {
-      throw (new exception(retCode.errorCode.LUIS_API_CALL_FAILED, text))
+      throw new Error(status.error.message)
     }
 
     return status
@@ -64,7 +69,7 @@ export class QnaBuildCore {
     const text = await response.text()
     const kb = JSON.parse(text)
     if (kb.error) {
-      throw (new exception(retCode.errorCode.LUIS_API_CALL_FAILED, text))
+      throw new Error(kb.error.message)
     }
 
     return kb
@@ -75,7 +80,7 @@ export class QnaBuildCore {
     const text = await response.text()
     const status = JSON.parse(text)
     if (status.error) {
-      throw (new exception(retCode.errorCode.LUIS_API_CALL_FAILED, text))
+      throw new Error(status.error.message)
     }
 
     return status
@@ -85,7 +90,7 @@ export class QnaBuildCore {
     const response = await this.service.createRequest(`/knowledgebases/${kbId}`, 'PUT', replaceKb)
     const text = await response.text()
     if (text) {
-      throw (new exception(retCode.errorCode.LUIS_API_CALL_FAILED, text))
+      throw new Error(JSON.parse(text).error.message)
     }
   }
 
@@ -93,7 +98,7 @@ export class QnaBuildCore {
     const response = await this.service.createRequest(`/knowledgebases/${kbId}`, 'POST')
     const text = await response.text()
     if (text) {
-      throw (new exception(retCode.errorCode.LUIS_API_CALL_FAILED, text))
+      throw new Error(JSON.parse(text).error.message)
     }
   }
 
@@ -101,7 +106,7 @@ export class QnaBuildCore {
     const response = await this.service.createRequest('/alterations', 'PUT', altJson)
     const text = await response.text()
     if (text) {
-      throw (new exception(retCode.errorCode.LUIS_API_CALL_FAILED, text))
+      throw new Error(JSON.parse(text).error.message)
     }
   }
 
@@ -110,7 +115,7 @@ export class QnaBuildCore {
     const text = await response.text()
     const endpointKeys = JSON.parse(text)
     if (endpointKeys.error) {
-      throw (new exception(retCode.errorCode.LUIS_API_CALL_FAILED, text))
+      throw new Error(endpointKeys.error.message)
     }
 
     return endpointKeys
@@ -120,7 +125,7 @@ export class QnaBuildCore {
     const response = await this.service.createRequest(`/knowledgebases/${kbId}`, 'DELETE')
     const text = await response.text()
     if (text) {
-      throw (new exception(retCode.errorCode.LUIS_API_CALL_FAILED, text))
+      throw new Error(JSON.parse(text).error.message)
     }
   }
 
