@@ -93,6 +93,28 @@ const helpers = {
         return splitReference.groups;
     },
     /**
+     * Helper function to parse link URIs in utterances
+     * @param {String} utterance
+     * @returns {Object} Object that contains luFile and ref. ref can be Intent-Name or ? or * or **
+     * @throws {exception} Throws on errors. exception object includes errCode and text. 
+     */
+    parseLinkURISync: function (utterance) {
+        let linkValueList = utterance.trim().match(new RegExp(/\(.*?\)/g));
+        let linkValue = linkValueList[0].replace('(', '').replace(')', '');
+        if (linkValue === '') throw (new exception(retCode.errorCode.INVALID_LU_FILE_REF, `[ERROR]: Invalid LU File Ref: "${utterance}"`));
+        // reference can either be #<Intent-Name> or #? or /*#? or /**#? or #*utterance* or #<Intent-Name>*patterns*
+        let splitRegExp = new RegExp(/^(?<fileName>.*?)(?<segment>#|\*+)(?<path>.*?)$/gim);
+        let splitReference = splitRegExp.exec(linkValue);
+        if (!splitReference) throw (new exception(retCode.errorCode.INVALID_LU_FILE_REF, `[ERROR]: Invalid LU File Ref: "${utterance}".\n Reference needs a qualifier - either a #Intent-Name or #? or *#? or **#? or #*utterances* etc.`));
+        if (splitReference.groups.segment.includes('*')) {
+            if (splitReference.groups.path === '') {
+                throw (new exception(retCode.errorCode.INVALID_LU_FILE_REF, `[ERROR]: Invalid LU File Ref: "${utterance}".\n '*' and '**' can only be used with QnA qualitifier. e.g. *#? and **#?`));
+            }
+            splitReference.groups.fileName += '*';
+        }
+        return splitReference.groups;
+    },
+    /**
      * Helper function to do a filter operation based search over an Array
      * @param {Array} srcList Object to filter on
      * @param {string} property Property to evaluate
@@ -131,9 +153,9 @@ const helpers = {
     isUtteranceLinkRef: function (utterance) {
         utterance = utterance || '';
         // Ensure only links are detected and passed on to be parsed.
-        // Valid link: [bar](xyz)
+        // Valid link: [bar](xyz), [bar][2]
         // Not a link: [bar](xyz|123), [bar[tar]](xyz), abc [foo](bar)
-        let linkDetectRegex = /^\[[^\[]+\]\([^|]+\)$/gi;
+        let linkDetectRegex = /^\[[^\[]+\](\([^|]+\)|\[[^\[]+\])$/gi;
         return linkDetectRegex.test(utterance);
     },
     /**
@@ -239,7 +261,11 @@ const updateToV7 = function(finalLUISJSON) {
                 }
             });
         });
-        (finalLUISJSON.entities || []).forEach(entity => transformAllEntityConstraintsToFeatures(entity));
+        let phraseListsNamesInFinal = [];
+        (finalLUISJSON.phraselists || []).forEach((item) => {
+            if (!phraseListsNamesInFinal.includes(item.name)) phraseListsNamesInFinal.push(item.name);
+        });
+        (finalLUISJSON.entities || []).forEach(entity => transformAllEntityConstraintsToFeatures(entity, phraseListsNamesInFinal));
         (finalLUISJSON.intents || []).forEach(intent => addIsRequiredProperty(intent));
         // do we have nDepthEntities?
         let nDepthEntityExists = (finalLUISJSON.entities || []).find(x => x.children !== undefined && Array.isArray(x.children) && x.children.length !== 0);
@@ -489,8 +515,8 @@ const objectSortByStartPos = function (objectArray) {
     return ObjectByStartPos
 }
 
-const transformAllEntityConstraintsToFeatures = function(entity) {
-    addIsRequiredProperty(entity);
+const transformAllEntityConstraintsToFeatures = function(entity, phraseListsInFinal) {
+    addIsRequiredProperty(entity, phraseListsInFinal);
     if (entity.hasOwnProperty("instanceOf") && entity.instanceOf !== null) {
         if (entity.hasOwnProperty("features") && Array.isArray(entity.features)) {
             let featureFound = (entity.features || []).find(i => i.modelName == entity.instanceOf);
@@ -513,10 +539,15 @@ const transformAllEntityConstraintsToFeatures = function(entity) {
     entity.children.forEach(c => transformAllEntityConstraintsToFeatures(c));
 };
 
-const addIsRequiredProperty = function(item) {
+const addIsRequiredProperty = function(item, phraseListInFinal = []) {
     (item.features || []).forEach(feature => {
         if (feature.isRequired === undefined)
             feature.isRequired = false;
+        if (feature.modelName !== undefined && phraseListInFinal.includes(feature.modelName))
+        {
+            feature.featureName = feature.modelName;
+            delete feature.modelName;
+        }    
         delete feature.featureType;
         delete feature.modelType;
     });
