@@ -31,42 +31,62 @@ describe('utils/filehelper test', () => {
     })
 
     it('File helper correctly build cross train config object', function () {
-        const configContent = {
-            "./main/main.lu": {
+        const rawConfigObject = {
+            "main": {
                 "rootDialog": true,
                 "triggers": {
-                    "dia1_trigger": ["./dia1/dia1.lu", "./dia2/dia2.lu"]
+                    "dia1_trigger": ["dia1", "dia2"]
                 }
             },
-            "./dia2/dia2.lu": {
+            "dia2": {
                 "triggers": {
                     "dia3_trigger": "",
-                    "": "./dia4/dia4.lu"
+                    "": "dia4"
                 }
             },
-            "./main/main.fr-fr.lu": {
+            "main.fr-fr": {
                 "rootDialog": true,
                 "triggers": {
-                    "dia1_trigger": "./dia1/dia1.fr-fr.lu"
+                    "dia1_trigger": "dia1.fr-fr"
                 }
             }
         }
 
-        let configObject = fileHelper.getConfigObject({ id: path.join(__dirname, 'config.json'), content: JSON.stringify(configContent) }, '_Interruption', true)
-        assert.equal(configObject.rootIds[0].includes('main.lu'), true)
-        assert.equal(configObject.rootIds[1].includes('main.fr-fr.lu'), true)
+        let configObject = fileHelper.getConfigObject(rawConfigObject, '_Interruption', true)
+        assert.equal(configObject.rootIds[0].includes('main'), true)
+        assert.equal(configObject.rootIds[1].includes('main.fr-fr'), true)
 
-        const triggerRuleKeys = [...Object.keys(configObject.triggerRules)]
-        assert.equal(triggerRuleKeys[0].includes('main.lu'), true)
-        assert.equal(triggerRuleKeys[1].includes('dia2.lu'), true)
-        assert.equal(triggerRuleKeys[2].includes('main.fr-fr.lu'), true)
+        let triggerRuleKeys = [...Object.keys(configObject.triggerRules)]
+        assert.equal(triggerRuleKeys[0].includes('main'), true)
+        assert.equal(triggerRuleKeys[1].includes('dia2'), true)
+        assert.equal(triggerRuleKeys[2].includes('main.fr-fr'), true)
 
-        const triggerRuleValues = [...Object.values(configObject.triggerRules)]
-        assert.equal(triggerRuleValues[0]['dia1_trigger'][0].includes('dia1.lu'), true)
-        assert.equal(triggerRuleValues[0]['dia1_trigger'][1].includes('dia2.lu'), true)
+        let triggerRuleValues = [...Object.values(configObject.triggerRules)]
+        assert.equal(triggerRuleValues[0]['dia1_trigger'][0].includes('dia1'), true)
+        assert.equal(triggerRuleValues[0]['dia1_trigger'][1].includes('dia2'), true)
         assert.equal(triggerRuleValues[1]['dia3_trigger'][0], '')
-        assert.equal(triggerRuleValues[1][''][0].includes('dia4.lu'), true)
-        assert.equal(triggerRuleValues[2]['dia1_trigger'][0].includes('dia1.fr-fr.lu'), true)
+        assert.equal(triggerRuleValues[1][''][0].includes('dia4'), true)
+        assert.equal(triggerRuleValues[2]['dia1_trigger'][0].includes('dia1.fr-fr'), true)
+
+        assert.equal(configObject.intentName, '_Interruption')
+
+        assert.equal(configObject.verbose, true)
+
+        configObject = fileHelper.getConfigObject(rawConfigObject, '_Interruption', true)
+        assert.equal(configObject.rootIds[0], 'main')
+        assert.equal(configObject.rootIds[1], 'main.fr-fr')
+
+        triggerRuleKeys = [...Object.keys(configObject.triggerRules)]
+        assert.equal(triggerRuleKeys[0], 'main')
+        assert.equal(triggerRuleKeys[1], 'dia2')
+        assert.equal(triggerRuleKeys[2], 'main.fr-fr')
+
+        triggerRuleValues = [...Object.values(configObject.triggerRules)]
+        assert.equal(triggerRuleValues[0]['dia1_trigger'][0], 'dia1')
+        assert.equal(triggerRuleValues[0]['dia1_trigger'][1], 'dia2')
+        assert.equal(triggerRuleValues[1]['dia3_trigger'][0], '')
+        assert.equal(triggerRuleValues[1][''][0], 'dia4')
+        assert.equal(triggerRuleValues[2]['dia1_trigger'][0], 'dia1.fr-fr')
 
         assert.equal(configObject.intentName, '_Interruption')
 
