@@ -1,4 +1,4 @@
-import fetch from 'node-fetch'
+import axios from 'axios'
 
 let headers = {
   'Content-Type': 'application/json',
@@ -9,9 +9,16 @@ export default {
   async get(
     url: string,
     subscriptionKey: string) {
-    setSubscriptionKey(subscriptionKey)
-    const response = await fetch(url, {method: 'GET', headers})
-    return response.json()
+    const resp = await httpRequest(
+      subscriptionKey,
+      {
+        method: 'GET',
+        url,
+        headers
+      })
+
+    return resp.data
+
   },
 
   async post(
@@ -19,29 +26,65 @@ export default {
     subscriptionKey: string,
     body: any,
     extraHeaders = {}) {
-    setSubscriptionKey(subscriptionKey)
     headers = {...headers, ...extraHeaders}
-    const response = await fetch(url, {method: 'POST', headers, body: JSON.stringify(body)})
-    return response.json()
+
+    const resp = await httpRequest(
+      subscriptionKey,
+      {
+        method: 'POST',
+        url,
+        data: body,
+        headers
+      })
+
+    return resp.data
   },
 
   async put(
     url: string,
     subscriptionKey: string,
     body: any) {
-    setSubscriptionKey(subscriptionKey)
-    const response = await fetch(url, {method: 'PUT', headers, body: JSON.stringify(body)})
+    const resp = await httpRequest(
+      subscriptionKey,
+      {
+        method: 'PUT',
+        url,
+        data: body,
+        headers
+      })
 
-    return isJSON(response) ? response.json() : {code: 'Success'}
+    return isJSON(resp.data) ? resp.data : {code: 'Success'}
   },
 
   async delete(
     url: string,
     subscriptionKey: string) {
-    setSubscriptionKey(subscriptionKey)
-    const response = await fetch(url, {method: 'DELETE', headers})
-    return isJSON(response) ? response.json() : {code: 'Success'}
+    const resp = await httpRequest(
+      subscriptionKey,
+      {
+        method: 'DELETE',
+        url,
+        headers
+      })
+    return isJSON(resp.data) ? resp.data : {code: 'Success'}
   }
+}
+
+const httpRequest = async function (subscriptionKey: string, config: any) {
+  setSubscriptionKey(subscriptionKey)
+  try {
+    return await axios(config)
+  } catch (error) {
+    if (error.response) {
+      // The request was made and the server responded with a status code
+      // that falls out of the range of 2xx
+      throw Error(error.response.statusText)
+    } else {
+      // Something happened in setting up the request that triggered an Error
+      throw Error(error.message)
+    }
+  }
+
 }
 
 const setSubscriptionKey = function (subscriptionKey: string) {
