@@ -12,10 +12,6 @@ const file = require('@microsoft/bf-lu/lib/utils/filehelper')
 const fileExtEnum = require('@microsoft/bf-lu/lib/parser/utils/helpers').FileExtTypeEnum
 const Content = require('@microsoft/bf-lu').V2.LU
 const LUOptions = require('@microsoft/bf-lu/lib/parser/lu/luOptions')
-const Settings = require('@microsoft/bf-lu/lib/parser/lubuild/settings')
-const MultiLanguageRecognizer = require('@microsoft/bf-lu/lib/parser/lubuild/multi-language-recognizer')
-const Recognizer = require('@microsoft/bf-lu/lib/parser/lubuild/recognizer')
-const CrosstrainedRecognizer = require('@microsoft/bf-lu/lib/parser/lubuild/cross-trained-recognizer')
 const Builder = require('@microsoft/bf-lu/lib/parser/lubuild/builder').Builder
 const recognizerType = require('@microsoft/bf-lu/lib/parser/utils/enums/recognizertypes')
 const utils = require('../../utils/index')
@@ -24,27 +20,27 @@ export default class LuisBuild extends Command {
   static description = 'Build lu files to train and publish luis applications'
 
   static examples = [`
-    $ bf luis:build --in {INPUT_FILE_OR_FOLDER} --authoringKey {AUTHORING_KEY} --botName {BOT_NAME} --dialog multiLanguage
+    $ bf luis:build --in {INPUT_FILE_OR_FOLDER} --authoringKey {AUTHORING_KEY} --botName {BOT_NAME}
   `]
 
   static flags: flags.Input<any> = {
-    help: flags.help({char: 'h'}),
+    help: flags.help({char: 'h', description: 'luis:build command help'}),
     in: flags.string({char: 'i', description: 'Lu file or folder'}),
     authoringKey: flags.string({description: 'LUIS authoring key'}),
     botName: flags.string({description: 'Bot name'}),
     region: flags.string({description: 'LUIS authoring region [westus|westeurope|australiaeast]', default: 'westus'}),
-    out: flags.string({char: 'o', description: 'Output folder name to write out .dialog files. If not specified, application ids will be output to console'}),
+    out: flags.string({char: 'o', description: 'Output folder name to write out .dialog and settings files. If not specified, application setting will be output to console'}),
     defaultCulture: flags.string({description: 'Culture code for the content. Infer from .lu if available. Defaults to en-us'}),
     fallbackLocale: flags.string({description: 'Locale to be used at the fallback if no locale specific recognizer is found. Only valid if --out is set'}),
     suffix: flags.string({description: 'Environment name as a suffix identifier to include in LUIS app name. Defaults to current logged in user alias'}),
-    dialog: flags.string({description: 'Dialog recognizer type [multiLanguage|crosstrained]', default: 'multiLanguage'}),
-    force: flags.boolean({char: 'f', description: 'If --out flag is provided, overwrites relevant dialog file', default: false}),
+    dialog: flags.string({description: 'Dialog recognizer type [multiLanguage|crosstrained]. No dialog recognizers will be generated if not specified. Only valid if --out is set'}),
+    force: flags.boolean({char: 'f', description: 'If --out flag is provided with the path to an existing file, overwrites that file', default: false}),
     luConfig: flags.string({description: 'Path to config for lu build which can contain switches for arguments'}),
-    deleteOldVersion: flags.boolean({description: 'Delete old version of LUIS application after building new one.'}),
-    log: flags.boolean({description: 'Write out log messages to console', default: false}),
+    deleteOldVersion: flags.boolean({description: 'Deletes old version of LUIS application after building new one.'}),
+    log: flags.boolean({description: 'Writes out log messages to console', default: false}),
     endpoint: flags.string({description: 'Luis authoring endpoint for publishing'}),
     schema: flags.string({description: 'Defines $schema for generated .dialog files'}),
-    isStaging: flags.boolean({description: 'Publish luis application to staging slot if set. Default to production slot', default: false})
+    isStaging: flags.boolean({description: 'Publishes luis application to staging slot if set. Default to production slot', default: false}),
   }
 
   async run() {
@@ -106,10 +102,6 @@ export default class LuisBuild extends Command {
       })
 
       let luContents: any[] = []
-      let recognizers = new Map<string, any>()
-      let multiRecognizers = new Map<string, any>()
-      let settings: any
-      let crosstrainedRecognizers = new Map<string, any>()
 
       if ((inVal && inVal !== '') || files.length > 0) {
         if (log) this.log('Loading files...\n')
@@ -121,44 +113,74 @@ export default class LuisBuild extends Command {
         }
 
         // de-dupe the files list
-        files = [...new Set(files)]
+        files = Array.from(new Set(files))
 
         // load lu contents from lu files
         // load existing recognizers, multiRecogniers and settings or create default ones
-        const loadedResources = await builder.loadContents(files, defaultCulture, suffix, region, schema)
-        luContents = loadedResources.luContents
-        recognizers = loadedResources.recognizers
-        multiRecognizers = loadedResources.multiRecognizers
-        settings = loadedResources.settings
-        crosstrainedRecognizers = loadedResources.crosstrainedRecognizers
+        luContents = await builder.loadContents(files, {
+          culture: defaultCulture,
+          log
+        })
       } else {
         // load lu content from stdin and create default recognizer, multiRecognier and settings
         if (log) this.log('Load lu content from stdin\n')
         const content = new Content(flags.stdin, new LUOptions('stdin', true, defaultCulture, path.join(process.cwd(), 'stdin')))
         luContents.push(content)
-        multiRecognizers.set('stdin', new MultiLanguageRecognizer(path.join(process.cwd(), 'stdin.lu.dialog'), {}))
-        settings = new Settings(path.join(process.cwd(), `luis.settings.${suffix}.${region}.json`), {})
-        const recognizer = Recognizer.load(content.path, content.name, path.join(process.cwd(), `${content.name}.dialog`), settings.get('stdin'), {})
-        recognizers.set(content.name, recognizer)
-        crosstrainedRecognizers.set('stdin', new CrosstrainedRecognizer(path.join(process.cwd(), 'stdin.lu.qna.dialog'), {}))
       }
 
       // update or create and then train and publish luis applications based on loaded resources
       if (log) this.log('Handling applications...')
-      const dialogContents = await builder.build(luContents, recognizers, authoringKey, endpoint, botName, suffix, fallbackLocale, deleteOldVersion, isStaging, multiRecognizers, settings, crosstrainedRecognizers, dialog)
+
+      // LUIS support maximun 100 versions for an application, keepVersions default set to 100 if deleteOldVersion is not specified.
+      let keptVersionCount = 100
+      if (deleteOldVersion) keptVersionCount = 1
+
+      const settingsContent = await builder.build(luContents, authoringKey, botName, {
+        endpoint,
+        suffix,
+        region,
+        keptVersionCount,
+        isStaging,
+        schema
+      })
 
       // write dialog assets based on config
       if (out) {
         const outputFolder = path.resolve(out)
-        const writeDone = await builder.writeDialogAssets(dialogContents, force, outputFolder, luConfig)
+        if (dialog) {
+          const dialogContents = await builder.generateDialogs(luContents, {
+            fallbackLocale,
+            schema,
+            dialog
+          })
+
+          let writeDone = await builder.writeDialogAssets(dialogContents, {
+            force,
+            out: outputFolder,
+            luConfig
+          })
+
+          if (writeDone) {
+            this.log(`Successfully wrote .dialog files to ${outputFolder}\n`)
+          } else {
+            this.log(`No changes to the .dialog files in ${outputFolder}\n`)
+          }
+        }
+
+        let writeDone = await builder.writeDialogAssets(settingsContent, {
+          force,
+          out: outputFolder,
+          luConfig
+        })
+
         if (writeDone) {
-          this.log(`Successfully wrote .dialog files to ${outputFolder}\n`)
+          this.log(`Successfully wrote settings file to ${outputFolder}\n`)
         } else {
-          this.log(`No changes to the .dialog files in ${outputFolder}\n`)
+          this.log(`No changes to settings file in ${outputFolder}\n`)
         }
       } else {
-        this.log('The published application ids:')
-        this.log(JSON.parse(dialogContents[dialogContents.length - 1].content).luis)
+        this.log('The published application setting:')
+        this.log(JSON.stringify(JSON.parse(settingsContent[0].content).luis, null, 4))
       }
     } catch (error) {
       if (error instanceof exception) {
