@@ -5,11 +5,14 @@ const uuidv1 = require('uuid/v1')
 const nock = require('nock')
 
 const compareFiles = async function (file1: string, file2: string) {
-  let result = await fs.readFile(path.join(__dirname, file1))
+  let result: string = ''
+  if (await fs.pathExists(path.join(__dirname, file1))) {
+    result = await fs.readFile(path.join(__dirname, file1))
+  }
+  
   let fixtureFile = await fs.readFile(path.join(__dirname, file2))
   result = result.toString().replace(/\r\n/g, '\n')
   fixtureFile = fixtureFile.toString().replace(/\r\n/g, '\n')
-  expect(fixtureFile).to.deep.equal(result)
   return result === fixtureFile
 }
 
@@ -17,6 +20,7 @@ describe('luis:build cli parameters test', () => {
   test
     .stdout()
     .command(['luis:build', '--help'])
+    .exit(1)
     .it('should print the help contents when --help is passed as an argument', ctx => {
       expect(ctx.stdout).to.contain('Build lu files to train and publish luis applications')
     })
@@ -25,6 +29,7 @@ describe('luis:build cli parameters test', () => {
     .stdout()
     .stderr()
     .command(['luis:build', '--in', `${path.join(__dirname, './../../fixtures/testcases/lubuild')}`, '--botName', 'Contoso'])
+    .exit(1)
     .it('displays an error if any required input parameters are missing', ctx => {
       expect(ctx.stderr).to.contain('Missing LUIS authoring key. Please pass authoring key with --authoringKey flag')
     })
@@ -33,6 +38,7 @@ describe('luis:build cli parameters test', () => {
     .stdout()
     .stderr()
     .command(['luis:build', '--authoringKey', uuidv1(), '--botName', 'Contoso'])
+    .exit(1)
     .it('displays an error if any required input parameters are missing', ctx => {
       expect(ctx.stderr).to.contain('Missing input. Please use stdin or pass a file or folder location with --in flag')
     })
@@ -41,6 +47,7 @@ describe('luis:build cli parameters test', () => {
     .stdout()
     .stderr()
     .command(['luis:build', '--authoringKey', uuidv1(), '--in', `${path.join(__dirname, './../../fixtures/testcases/lubuild')}`])
+    .exit(1)
     .it('displays an error if any required input parameters are missing', ctx => {
       expect(ctx.stderr).to.contain('Missing bot name. Please pass bot name with --botName flag')
     })
@@ -49,6 +56,7 @@ describe('luis:build cli parameters test', () => {
     .stdout()
     .stderr()
     .command(['luis:build', '--authoringKey', uuidv1(), '--in', `${path.join(__dirname, './../../fixtures/testcases/lubuild/file-name-duplicated')}`, '--botName', 'Contoso'])
+    .exit(1)
     .it('displays an error if files with same name and locale are found', ctx => {
       expect(ctx.stderr).to.contain('Files with same name and locale are found')
     })
@@ -57,6 +65,7 @@ describe('luis:build cli parameters test', () => {
     .stdout()
     .stderr()
     .command(['luis:build', '--authoringKey', uuidv1(), '--in', `${path.join(__dirname, './../../fixtures/testcases/invalid_import_file.lu')}`, '--botName', 'Contoso'])
+    .exit(1)
     .it('displays an error if error occurs in parsing lu content', ctx => {
       expect(ctx.stderr).to.contain('Invalid LU file')
       expect(ctx.stderr).to.contain('bad3.lu')
@@ -67,6 +76,7 @@ describe('luis:build cli parameters test', () => {
     .stdout()
     .stderr()
     .command(['luis:build', '--authoringKey', uuidv1(), '--in', `${path.join(__dirname, './../../fixtures/testcases/lubuild')}`, '--botName', 'Contoso', '--dialog', 'cross-train'])
+    .exit(1)
     .it('displays an error if option specified by --dialog is not right', ctx => {
       expect(ctx.stderr).to.contain('Recognizer type specified by --dialog is not right. Please specify [multiLanguage|crosstrained]')
     })
@@ -153,6 +163,16 @@ describe('luis:build update application succeed when utterances changed', () => 
       .reply(201, '0.2')
 
     nock('https://westus.api.cognitive.microsoft.com')
+      .get(uri => uri.includes('apps'))
+      .reply(200, [
+        {
+          version: '0.2'
+        },
+        {
+          version: '0.1'
+        }])
+
+    nock('https://westus.api.cognitive.microsoft.com')
       .post(uri => uri.includes('train'))
       .reply(202, {
         statusId: 2,
@@ -216,6 +236,16 @@ describe('luis:build update application succeed when utterances added', () => {
     nock('https://westus.api.cognitive.microsoft.com')
       .post(uri => uri.includes('import'))
       .reply(201, '0.2')
+    
+    nock('https://westus.api.cognitive.microsoft.com')
+      .get(uri => uri.includes('apps'))
+      .reply(200, [
+        {
+          version: '0.2'
+        },
+        {
+          version: '0.1'
+        }])
 
     nock('https://westus.api.cognitive.microsoft.com')
       .post(uri => uri.includes('train'))
@@ -288,7 +318,39 @@ describe('luis:build not update application if no changes', () => {
     })
 })
 
-describe('luis:build write dialog assets successfully if --dialog set to multiLanguage', () => {
+describe('luis:build not create application if kb name id only case different', () => {
+  const existingLuisApp = require('./../../fixtures/testcases/lubuild/sandwich/luis/test(development)-sandwich.en-us.lu.json')
+  before(function () {
+    nock('https://westus.api.cognitive.microsoft.com')
+      .get(uri => uri.includes('apps'))
+      .reply(200, [{
+        name: 'test(development)-sandwich.en-us.lu',
+        id: 'f8c64e2a-8635-3a09-8f78-39d7adc76ec5'
+      }])
+
+    nock('https://westus.api.cognitive.microsoft.com')
+      .get(uri => uri.includes('apps'))
+      .reply(200, {
+        name: 'test(development)-sandwich.en-us.lu',
+        id: 'f8c64e2a-8635-3a09-8f78-39d7adc76ec5',
+        activeVersion: '0.1'
+      })
+
+    nock('https://westus.api.cognitive.microsoft.com')
+      .get(uri => uri.includes('export'))
+      .reply(200, existingLuisApp)
+  })
+
+  test
+    .stdout()
+    .command(['luis:build', '--in', './test/fixtures/testcases/lubuild/sandwich/lufiles/sandwich.en-us.lu', '--authoringKey', uuidv1(), '--botName', 'TEST', '--log', '--suffix', 'development'])
+    .it('should not create a luis application when an app with only case different name already existed in service', ctx => {
+      expect(ctx.stdout).to.contain('Handling applications...')
+      expect(ctx.stdout).to.contain('no changes')
+    })
+})
+
+describe('luis:build write dialog and settings assets successfully if --dialog set to multiLanguage', () => {
   const existingLuisApp = require('./../../fixtures/testcases/lubuild/sandwich/luis/test(development)-sandwich.en-us.lu.json')
   before(async function () {
     await fs.ensureDir(path.join(__dirname, './../../../results/'))
@@ -319,15 +381,15 @@ describe('luis:build write dialog assets successfully if --dialog set to multiLa
 
   test
     .stdout()
-    .command(['luis:build', '--in', './test/fixtures/testcases/lubuild/sandwich/lufiles/sandwich.en-us.lu', '--authoringKey', uuidv1(), '--botName', 'test', '--out', './results', '--log', '--suffix', 'development'])
-    .it('should write dialog assets successfully when --dialog set to multiLanguage', async ctx => {
+    .command(['luis:build', '--in', './test/fixtures/testcases/lubuild/sandwich/lufiles/sandwich.en-us.lu', '--authoringKey', uuidv1(), '--botName', 'test', '--out', './results', '--dialog', 'multiLanguage', '--log', '--suffix', 'development'])
+    .it('should write dialog and settings assets successfully when --dialog set to multiLanguage', async ctx => {
       expect(await compareFiles('./../../../results/luis.settings.development.westus.json', './../../fixtures/testcases/lubuild/sandwich/config/luis.settings.development.westus.json')).to.be.true
       expect(await compareFiles('./../../../results/sandwich.en-us.lu.dialog', './../../fixtures/testcases/lubuild/sandwich/dialogs/sandwich.en-us.lu.dialog')).to.be.true
       expect(await compareFiles('./../../../results/sandwich.lu.dialog', './../../fixtures/testcases/lubuild/sandwich/dialogs/sandwich.lu.dialog')).to.be.true
     })
 })
 
-describe('luis:build write dialog assets successfully if --dialog set to crosstrained', () => {
+describe('luis:build write dialog and settings assets successfully if --dialog set to crosstrained', () => {
   const existingLuisApp = require('./../../fixtures/testcases/lubuild/sandwich/luis/test(development)-sandwich.en-us.lu.json')
   before(async function () {
     await fs.ensureDir(path.join(__dirname, './../../../results/'))
@@ -359,14 +421,14 @@ describe('luis:build write dialog assets successfully if --dialog set to crosstr
   test
     .stdout()
     .command(['luis:build', '--in', './test/fixtures/testcases/lubuild/sandwich/lufiles/sandwich.en-us.lu', '--authoringKey', uuidv1(), '--botName', 'test', '--dialog', 'crosstrained', '--out', './results', '--log', '--suffix', 'development'])
-    .it('should write dialog assets successfully when --dialog set to crosstrained', async ctx => {
+    .it('should write dialog and settings assets successfully when --dialog set to crosstrained', async ctx => {
       expect(await compareFiles('./../../../results/luis.settings.development.westus.json', './../../fixtures/testcases/lubuild/sandwich/config/luis.settings.development.westus.json')).to.be.true
       expect(await compareFiles('./../../../results/sandwich.lu.qna.dialog', './../../fixtures/testcases/lubuild/sandwich/dialogs/sandwich.lu.qna.dialog')).to.be.true
       expect(await compareFiles('./../../../results/sandwich.lu.dialog', './../../fixtures/testcases/lubuild/sandwich/dialogs/sandwich.lu.dialog')).to.be.true
     })
 })
 
-describe('luis:build write crosstrained recognizer asset successfully if lu file is empty and --dialog set to crosstrained', () => {
+describe('luis:build write crosstrained recognizer asset and settings successfully if lu file is empty and --dialog set to crosstrained', () => {
   before(async function () {
     await fs.ensureDir(path.join(__dirname, './../../../results/'))
 
@@ -502,7 +564,7 @@ describe('luis:build create multiple applications successfully when input is a f
   test
     .stdout()
     .command(['luis:build', '--in', './test/fixtures/testcases/lubuild/foo/lufiles', '--authoringKey', uuidv1(), '--botName', 'test', '--dialog', 'multiLanguage', '--out', './results', '--log', '--suffix', 'development'])
-    .it('should create multiple applications and write dialog assets successfully when input is a folder', async ctx => {
+    .it('should create multiple applications and write dialog and settings assets successfully when input is a folder', async ctx => {
       expect(ctx.stdout).to.contain('foo.fr-fr.lu loaded')
       expect(ctx.stdout).to.contain('foo.lu loaded')
       expect(ctx.stdout).to.contain('foo.zh-cn.lu loaded')
@@ -522,121 +584,6 @@ describe('luis:build create multiple applications successfully when input is a f
       expect(await compareFiles('./../../../results/foo.en-us.lu.dialog', './../../fixtures/testcases/lubuild/foo/dialogs/foo.en-us.lu.dialog')).to.be.true
       expect(await compareFiles('./../../../results/foo.fr-fr.lu.dialog', './../../fixtures/testcases/lubuild/foo/dialogs/foo.fr-fr.lu.dialog')).to.be.true
       expect(await compareFiles('./../../../results/foo.zh-cn.lu.dialog', './../../fixtures/testcases/lubuild/foo/dialogs/foo.zh-cn.lu.dialog')).to.be.true
-    })
-})
-
-describe('luis:build update dialog assets successfully when dialog assets exist', () => {
-  const existingLuisApp_EN_US = require('./../../fixtures/testcases/lubuild/foo2/luis/test(development)-foo.en-us.lu.json')
-  const existingLuisApp_FR_FR = require('./../../fixtures/testcases/lubuild/foo2/luis/test(development)-foo.fr-fr.lu.json')
-  before(async function () {
-    await fs.ensureDir(path.join(__dirname, './../../../results/'))
-
-    nock('https://westus.api.cognitive.microsoft.com')
-      .get(uri => uri.includes('apps'))
-      .reply(200, [
-        {
-          name: 'test(development)-foo.fr-fr.lu',
-          id: 'f8c64e2a-1111-3a09-8f78-39d7adc76ec5'
-        },
-        {
-          name: 'test(development)-foo.en-us.lu',
-          id: 'f8c64e2a-2222-3a09-8f78-39d7adc76ec5'
-        }
-      ])
-
-    nock('https://westus.api.cognitive.microsoft.com')
-      .post(uri => uri.includes('import'))
-      .reply(201, {
-        appId: 'f8c64e2a-3333-3a09-8f78-39d7adc76ec5'
-      })
-
-    nock('https://westus.api.cognitive.microsoft.com')
-      .get(uri => uri.includes('apps') && uri.includes('f8c64e2a-2222-3a09-8f78-39d7adc76ec5'))
-      .reply(200, {
-        name: 'test(development)-foo.en-us.lu',
-        id: 'f8c64e2a-2222-3a09-8f78-39d7adc76ec5',
-        activeVersion: '0.1'
-      })
-
-    nock('https://westus.api.cognitive.microsoft.com')
-      .get(uri => uri.includes('apps') && uri.includes('f8c64e2a-1111-3a09-8f78-39d7adc76ec5'))
-      .reply(200, {
-        name: 'test(development)-foo.fr-fr.lu',
-        id: 'f8c64e2a-1111-3a09-8f78-39d7adc76ec5',
-        activeVersion: '0.1'
-      })
-
-    nock('https://westus.api.cognitive.microsoft.com')
-      .post(uri => uri.includes('train'))
-      .reply(202, {
-        statusId: 2,
-        status: 'UpToDate'
-      })
-
-    nock('https://westus.api.cognitive.microsoft.com')
-      .get(uri => uri.includes('export') && uri.includes('f8c64e2a-2222-3a09-8f78-39d7adc76ec5'))
-      .reply(200, existingLuisApp_EN_US)
-
-    nock('https://westus.api.cognitive.microsoft.com')
-      .get(uri => uri.includes('export') && uri.includes('f8c64e2a-1111-3a09-8f78-39d7adc76ec5'))
-      .reply(200, existingLuisApp_FR_FR)
-
-
-    nock('https://westus.api.cognitive.microsoft.com')
-      .get(uri => uri.includes('train'))
-      .reply(200, [{
-        modelId: '99999',
-        details: {
-          statusId: 0,
-          status: 'Success',
-          exampleCount: 0
-        }
-      }])
-
-    nock('https://westus.api.cognitive.microsoft.com')
-      .post(uri => uri.includes('publish'))
-      .reply(201, {
-        versionId: '0.2',
-        isStaging: false
-      })
-  })
-
-  after(async function () {
-    await fs.remove(path.join(__dirname, './../../../results/'))
-  })
-
-  test
-    .stdout()
-    .command(['luis:build', '--in', './test/fixtures/testcases/lubuild/foo2/lufiles-and-dialog-assets', '--authoringKey', uuidv1(), '--botName', 'test', '--dialog', 'multiLanguage', '--out', './results', '--log', '--suffix', 'development'])
-    .it('should update dialog assets successfully when dialog assets exist', async ctx => {
-      expect(ctx.stdout).to.contain('foo.fr-fr.lu loaded')
-      expect(ctx.stdout).to.contain('foo.lu loaded')
-      expect(ctx.stdout).to.contain('foo.zh-cn.lu loaded')
-
-      expect(ctx.stdout).to.contain('luis.settings.development.westus.json loaded')
-      expect(ctx.stdout).to.contain('foo.lu.dialog loaded')
-
-      expect(ctx.stdout).to.contain('Handling applications...')
-
-      expect(ctx.stdout).to.contain('foo.en-us.lu.dialog loaded')
-      expect(ctx.stdout).to.contain('foo.fr-fr.lu.dialog loaded')
-
-      expect(ctx.stdout).to.contain('[WARN]: empty intent(s) # emptyIntent are filtered when handling luis application')
-
-      expect(ctx.stdout).to.contain('Creating LUIS.ai application')
-      expect(ctx.stdout).to.contain('training version=0.1')
-      expect(ctx.stdout).to.contain('waiting for training for version=0.1')
-      expect(ctx.stdout).to.contain('publishing version=0.1')
-
-      expect(ctx.stdout).to.contain('foo.fr-fr.lu no changes')
-      expect(ctx.stdout).to.contain('foo.lu no changes')
-      expect(ctx.stdout).to.contain('foo.zh-cn.lu publishing finished')
-
-      expect(await compareFiles('./../../../results/luis.settings.development.westus.json', './../../fixtures/testcases/lubuild/foo2/config/luis.settings.development.westus.json')).to.be.true
-      expect(await compareFiles('./../../../results/foo.lu.dialog', './../../fixtures/testcases/lubuild/foo2/dialogs/foo.lu.dialog')).to.be.true
-      expect(await compareFiles('./../../../results/foo.en-us.lu.dialog', './../../fixtures/testcases/lubuild/foo2/dialogs/foo.en-us.lu.dialog')).to.be.true
-      expect(await compareFiles('./../../../results/foo.fr-fr.lu.dialog', './../../fixtures/testcases/lubuild/foo2/dialogs/foo.fr-fr.lu.dialog')).to.be.true
-      expect(await compareFiles('./../../../results/foo.zh-cn.lu.dialog', './../../fixtures/testcases/lubuild/foo2/dialogs/foo.zh-cn.lu.dialog')).to.be.true
     })
 })
 
@@ -704,15 +651,29 @@ describe('luis:build update application succeed with parameters set from luconfi
       .get(uri => uri.includes('apps'))
       .reply(200, [
         {
-          version: '0.1'
+          version: '0.4'
+        },
+        {
+          version: '0.3'
         },
         {
           version: '0.2'
+        },
+        {
+          version: '0.1'
         }
       ])
     
     nock('https://chinaeast2.api.cognitive.azure.cn')
-      .delete(uri => uri.includes('apps'))
+      .delete(uri => uri.includes('0.1'))
+      .reply(200)
+
+    nock('https://chinaeast2.api.cognitive.azure.cn')
+      .delete(uri => uri.includes('0.2'))
+      .reply(200)
+
+    nock('https://chinaeast2.api.cognitive.azure.cn')
+      .delete(uri => uri.includes('0.3'))
       .reply(200)
 
     nock('https://chinaeast2.api.cognitive.azure.cn')
@@ -850,6 +811,16 @@ describe('luis:build update application succeed when activeVersion is null', () 
       .reply(201, '0.2')
 
     nock('https://westus.api.cognitive.microsoft.com')
+      .get(uri => uri.includes('apps'))
+      .reply(200, [
+        {
+          version: '0.2'
+        },
+        {
+          version: '0.1'
+        }])
+
+    nock('https://westus.api.cognitive.microsoft.com')
       .post(uri => uri.includes('train'))
       .reply(202, {
         statusId: 2,
@@ -942,7 +913,7 @@ describe('luis:build create a new application successfully with endpoint overrid
     })
 })
 
-describe('luis:build write dialog assets successfully if schema is specified', () => {
+describe('luis:build write dialog and settings assets successfully if schema is specified', () => {
   const existingLuisApp = require('./../../fixtures/testcases/lubuild/sandwich/luis/test(development)-sandwich.en-us.lu.json')
   before(async function () {
     await fs.ensureDir(path.join(__dirname, './../../../results/'))
@@ -974,7 +945,7 @@ describe('luis:build write dialog assets successfully if schema is specified', (
   test
     .stdout()
     .command(['luis:build', '--in', './test/fixtures/testcases/lubuild/sandwich/lufiles/sandwich.en-us.lu', '--authoringKey', uuidv1(), '--botName', 'test', '--out', './results', '--log', '--suffix', 'development', '--dialog', 'crosstrained', '--schema', 'https://raw.githubusercontent.com/microsoft/BotFramework-Composer/stable/Composer/packages/server/schemas/sdk.schema'])
-    .it('should write dialog assets successfully if schema is specified', async ctx => {
+    .it('should write dialog and settings assets successfully if schema is specified', async ctx => {
       expect(await compareFiles('./../../../results/luis.settings.development.westus.json', './../../fixtures/testcases/lubuild/sandwich/config/luis.settings.development.westus.json')).to.be.true
       expect(await compareFiles('./../../../results/sandwich.en-us.lu.dialog', './../../fixtures/testcases/lubuild/sandwich/dialogs-with-schema/sandwich.en-us.lu.dialog')).to.be.true
       expect(await compareFiles('./../../../results/sandwich.lu.dialog', './../../fixtures/testcases/lubuild/sandwich/dialogs-with-schema/sandwich.lu.dialog')).to.be.true
@@ -982,7 +953,7 @@ describe('luis:build write dialog assets successfully if schema is specified', (
     })
 })
 
-describe('luis:build write dialog assets successfully when empty files exist', () => {
+describe('luis:build write dialog and settings assets successfully when empty files exist', () => {
   before(async function () {
     await fs.ensureDir(path.join(__dirname, './../../../results/'))
 
@@ -1032,7 +1003,7 @@ describe('luis:build write dialog assets successfully when empty files exist', (
   test
     .stdout()
     .command(['luis:build', '--in', './test/fixtures/testcases/lubuild/empty-file/lufiles', '--authoringKey', uuidv1(), '--botName', 'test', '--dialog', 'crosstrained', '--out', './results', '--log', '--suffix', 'development'])
-    .it('should write dialog assets successfully when empty files exist', async ctx => {
+    .it('should write dialog and settings assets successfully when empty files exist', async ctx => {
       expect(ctx.stdout).to.contain('empty.lu loaded')
       expect(ctx.stdout).to.contain('non-empty.lu loaded')
 
@@ -1049,5 +1020,66 @@ describe('luis:build write dialog assets successfully when empty files exist', (
       expect(await compareFiles('./../../../results/non-empty.lu.dialog', './../../fixtures/testcases/lubuild/empty-file/dialogs/non-empty.lu.dialog')).to.be.true
       expect(await compareFiles('./../../../results/non-empty.en-us.lu.dialog', './../../fixtures/testcases/lubuild/empty-file/dialogs/non-empty.en-us.lu.dialog')).to.be.true
       expect(await compareFiles('./../../../results/non-empty.lu.qna.dialog', './../../fixtures/testcases/lubuild/empty-file/dialogs/non-empty.lu.qna.dialog')).to.be.true
+    })
+})
+
+describe('luis:build write settings assets only successfully if --dialog is not set', () => {
+  const existingLuisApp = require('./../../fixtures/testcases/lubuild/sandwich/luis/test(development)-sandwich.en-us.lu.json')
+  before(async function () {
+    await fs.ensureDir(path.join(__dirname, './../../../results/'))
+
+    nock('https://westus.api.cognitive.microsoft.com')
+      .get(uri => uri.includes('apps'))
+      .reply(200, [{
+        name: 'test(development)-sandwich.en-us.lu',
+        id: 'f8c64e2a-8635-3a09-8f78-39d7adc76ec5'
+      }])
+
+    nock('https://westus.api.cognitive.microsoft.com')
+      .get(uri => uri.includes('apps'))
+      .reply(200, {
+        name: 'test(development)-sandwich.en-us.lu',
+        id: 'f8c64e2a-8635-3a09-8f78-39d7adc76ec5',
+        activeVersion: '0.1'
+      })
+
+    nock('https://westus.api.cognitive.microsoft.com')
+      .get(uri => uri.includes('export'))
+      .reply(200, existingLuisApp)
+  })
+
+  after(async function () {
+    await fs.remove(path.join(__dirname, './../../../results/'))
+  })
+
+  test
+    .stdout()
+    .command(['luis:build', '--in', './test/fixtures/testcases/lubuild/sandwich/lufiles/sandwich.en-us.lu', '--authoringKey', uuidv1(), '--botName', 'test', '--out', './results', '--log', '--suffix', 'development'])
+    .it('should write settings assets only successfully if --dialog is not set', async ctx => {
+      expect(await compareFiles('./../../../results/luis.settings.development.westus.json', './../../fixtures/testcases/lubuild/sandwich/config/luis.settings.development.westus.json')).to.be.true
+      expect(await compareFiles('./../../../results/sandwich.lu.qna.dialog', './../../fixtures/testcases/lubuild/sandwich/dialogs/sandwich.lu.qna.dialog')).to.be.false
+      expect(await compareFiles('./../../../results/sandwich.lu.dialog', './../../fixtures/testcases/lubuild/sandwich/dialogs/sandwich.lu.dialog')).to.be.false
+    })
+})
+
+describe('luis:build throw luis build failed exception successfully', () => {
+  before(function () {
+    nock('https://westus.api.cognitive.microsoft.com')
+      .get(uri => uri.includes('apps'))
+      .reply(401, {
+        error: {
+          code: 401,
+          message: 'Access denied due to invalid subscription key.'
+        }
+      })
+  })
+
+  test
+    .stdout()
+    .stderr()
+    .command(['luis:build', '--in', './test/fixtures/testcases/lubuild/sandwich//lufiles/sandwich.en-us.lu', '--authoringKey', uuidv1(), '--botName', 'test', '--log', '--suffix', 'development'])
+    .exit(1)
+    .it('should throw luis build failed exception successfully', ctx => {
+      expect(ctx.stderr).to.contain('Luis build failed: Access denied due to invalid subscription key.')
     })
 })
