@@ -1,9 +1,8 @@
+import sinon from 'sinon'
+import uuidv1 from 'uuid/v1'
 import {expect, test} from '@oclif/test'
-const sinon = require('sinon')
-const uuidv1 = require('uuid/v1')
+
 const utils = require('../../../../src/utils/index')
-const fs = require('fs-extra')
-import * as rimraf from 'rimraf'
 
 describe('luis:train:run', () => {
 
@@ -18,6 +17,7 @@ describe('luis:train:run', () => {
   test
   .stdout()
   .command(['luis:train:run', '--help'])
+  .exit(1)
   .it('should print the help contents when --help is passed as an argument', ctx => {
     expect(ctx.stdout).to.contain('Issues asynchronous training request for LUIS application')
   })
