@@ -1,9 +1,8 @@
+import sinon from 'sinon'
+import uuidv1 from 'uuid/v1'
 import {expect, test} from '@oclif/test'
-const sinon = require('sinon')
-const uuidv1 = require('uuid/v1')
+
 const utils = require('../../../../src/utils/index')
-const fs = require('fs-extra')
-import * as rimraf from 'rimraf'
 
 describe('luis:train:show', () => {
 
@@ -18,6 +17,7 @@ describe('luis:train:show', () => {
   test
   .stdout()
   .command(['luis:train:show', '--help'])
+  .exit(1)
   .it('should print the help contents when --help is passed as an argument', ctx => {
     expect(ctx.stdout).to.contain('Shows training status')
   })
@@ -26,6 +26,7 @@ describe('luis:train:show', () => {
   .stdout()
   .stderr()
   .command(['luis:train:show', '--appId', uuidv1(), '--endpoint', 'https://westus.api.cognitive.microsoft.com',  '--versionId', '0.1'])
+  .exit(1)
   .it('displays an error if any required input parameters are missing', ctx => {
     expect(ctx.stderr).to.contain(`Required input property 'subscriptionKey' missing.`)
   })
