@@ -16,6 +16,7 @@ describe('luis:version:rename', () => {
   test
   .stdout()
   .command(['luis:version:rename', '--help'])
+  .exit(1)
   .it('should print the help contents when --help is passed as an argument', ctx => {
     expect(ctx.stdout).to.contain('Renames application version')
   })
@@ -24,6 +25,7 @@ describe('luis:version:rename', () => {
   .stdout()
   .stderr()
   .command(['luis:version:rename', '--endpoint', 'https://westus.api.cognitive.microsoft.com', '--appId', uuidv1(), '--subscriptionKey', uuidv1(), '--newVersionId', '0.2'])
+  .exit(1)
   .it('displays an error if any required input parameters are missing', ctx => {
     expect(ctx.stderr).to.contain(`Required input property 'versionId' missing.`)
   })
@@ -32,6 +34,7 @@ describe('luis:version:rename', () => {
   .stdout()
   .stderr()
   .command(['luis:version:rename', '--endpoint', 'https://westus.api.cognitive.microsoft.com', '--appId', uuidv1(), '--versionId', '0.1', '--newVersionId', '0.2'])
+  .exit(1)
   .it('displays an error if any required input parameters are missing', ctx => {
     expect(ctx.stderr).to.contain(`Required input property 'subscriptionKey' missing.`)
   })
@@ -51,8 +54,9 @@ describe('luis:version:rename', () => {
   .stdout()
   .stderr()
   .command(['luis:version:rename', '--endpoint', 'undefined', '--subscriptionKey', uuidv1(), '--appId', uuidv1(), '--versionId', '0.1', '--newVersionId', '0.2'])
+  .exit(1)
   .it('fails to rename application version and displays an error message if the endpoint is undefined', ctx => {
-    expect(ctx.stderr).to.contain('Failed to rename app version: TypeError: Only absolute URLs are supported\n')
+    expect(ctx.stderr).to.contain('Failed to rename app version: Error')
   })
 
 })
