@@ -29,7 +29,7 @@ export async function processInputs(flags: any, payload: any, configfile: string
       result.requestBody = input ? await file.getContentFromFile(input) : stdin
       try {
         result.requestBody = JSON.parse(result.requestBody)
-      } catch(ex) {
+      } catch (ex) {
         flags.qnaFormat = true
       }
     }
@@ -81,11 +81,11 @@ export async function processFlags(flags: any, flagLabels: string[], configDir: 
 
       input[flag] = flags[flag] || (config ? config[configPrefix + flag] : null)
     })
-  
+
   return input
 }
 
-async function getUserConfig (configPath: string) {
+async function getUserConfig(configPath: string) {
   if (fs.existsSync(path.join(configPath, 'config.json'))) {
     return fs.readJSON(path.join(configPath, 'config.json'), {throws: false})
   }
@@ -93,20 +93,20 @@ async function getUserConfig (configPath: string) {
   return {}
 }
 
-function filterByAllowedConfigValues (configObj: any, prefix: string) {
+function filterByAllowedConfigValues(configObj: any, prefix: string) {
   const allowedConfigValues = [`${prefix}kbId`, `${prefix}endpoint`, `${prefix}region`, `${prefix}subscriptionKey`]
   const filtered = Object.keys(configObj)
   .filter(key => allowedConfigValues.includes(key))
   .reduce((filteredConfigObj: any, key) => {
     filteredConfigObj[key] = configObj[key]
-    
+
     return filteredConfigObj
   }, {})
 
   return filtered
 }
 
-function filterConfig (config: any, prefix: string) {
+function filterConfig(config: any, prefix: string) {
   return Object.keys(config)
     .filter((key: string) => key.startsWith(prefix))
     .reduce((filteredConfig: any, key: string) => {
