@@ -5,11 +5,14 @@ const uuidv1 = require('uuid/v1')
 const nock = require('nock')
 
 const compareFiles = async function (file1: string, file2: string) {
-  let result = await fs.readFile(path.join(__dirname, file1))
+  let result: string = ''
+  if (await fs.pathExists(path.join(__dirname, file1))) {
+    result = await fs.readFile(path.join(__dirname, file1))
+  }
+
   let fixtureFile = await fs.readFile(path.join(__dirname, file2))
   result = result.toString().replace(/\r\n/g, '\n')
   fixtureFile = fixtureFile.toString().replace(/\r\n/g, '\n')
-  expect(fixtureFile).to.deep.equal(result)
   return result === fixtureFile
 }
 
@@ -17,6 +20,7 @@ describe('qnamaker:build cli parameters test', () => {
   test
     .stdout()
     .command(['qnamaker:build', '--help'])
+    .exit(1)
     .it('should print the help contents when --help is passed as an argument', ctx => {
       expect(ctx.stdout).to.contain('Build .qna files to create or update qnamaker knowledge bases and qnamaker alterations')
     })
@@ -25,6 +29,7 @@ describe('qnamaker:build cli parameters test', () => {
     .stdout()
     .stderr()
     .command(['qnamaker:build', '--in', `${path.join(__dirname, './../../fixtures/testcases/qnabuild')}`, '--botName', 'Contoso'])
+    .exit(1)
     .it('displays an error if any required input parameters are missing', ctx => {
       expect(ctx.stderr).to.contain('Missing qnamaker subscription key. Please pass subscription key with --subscriptionKey flag or specify via bf config:set:qnamaker --subscriptionKey.')
     })
@@ -33,6 +38,7 @@ describe('qnamaker:build cli parameters test', () => {
     .stdout()
     .stderr()
     .command(['qnamaker:build', '--subscriptionKey', uuidv1(), '--botName', 'Contoso'])
+    .exit(1)
     .it('displays an error if any required input parameters are missing', ctx => {
       expect(ctx.stderr).to.contain('Missing input. Please use stdin or pass a file or folder location with --in flag')
     })
@@ -41,6 +47,7 @@ describe('qnamaker:build cli parameters test', () => {
     .stdout()
     .stderr()
     .command(['qnamaker:build', '--subscriptionKey', uuidv1(), '--in', `${path.join(__dirname, './../../fixtures/testcases/qnabuild')}`])
+    .exit(1)
     .it('displays an error if any required input parameters are missing', ctx => {
       expect(ctx.stderr).to.contain('Missing bot name. Please pass bot name with --botName flag or specify via --qnaConfig.')
     })
@@ -49,6 +56,7 @@ describe('qnamaker:build cli parameters test', () => {
     .stdout()
     .stderr()
     .command(['qnamaker:build', '--subscriptionKey', uuidv1(), '--in', `${path.join(__dirname, './../../fixtures/testcases/qnabuild')}`, '--botName', 'Contoso', '--dialog', 'cross-train'])
+    .exit(1)
     .it('displays an error if option specified by --dialog is not right', ctx => {
       expect(ctx.stderr).to.contain('Recognizer type specified by --dialog is not right. Please specify [multiLanguage|crosstrained]')
     })
@@ -216,7 +224,7 @@ describe('qnamaker:build not update knowledge if no changes', () => {
     })
 })
 
-describe('qnamaker:build write dialog assets successfully if --dialog set to multiLanguage', () => {
+describe('qnamaker:build write dialog and settings assets successfully if --dialog set to multiLanguage', () => {
   before(async function () {
     await fs.ensureDir(path.join(__dirname, './../../../results/'))
 
@@ -257,8 +265,8 @@ describe('qnamaker:build write dialog assets successfully if --dialog set to mul
 
   test
     .stdout()
-    .command(['qnamaker:build', '--in', './test/fixtures/testcases/qnabuild/sandwich/qnafiles/sandwich2.en-us.qna', '--subscriptionKey', uuidv1(), '--botName', 'test', '--out', './results', '--log', '--suffix', 'development'])
-    .it('should write dialog assets successfully when --dialog set to multiLanguage', async ctx => {
+    .command(['qnamaker:build', '--in', './test/fixtures/testcases/qnabuild/sandwich/qnafiles/sandwich2.en-us.qna', '--subscriptionKey', uuidv1(), '--botName', 'test', '--out', './results', '--dialog', 'multiLanguage', '--log', '--suffix', 'development'])
+    .it('should write dialog and settings assets successfully when --dialog set to multiLanguage', async ctx => {
       expect(ctx.stdout).to.contain('yyyyyyyyyyyyyyyyyyyyyyyyyyyyyyy')
       expect(ctx.stdout).to.contain('xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx')
       
@@ -268,7 +276,7 @@ describe('qnamaker:build write dialog assets successfully if --dialog set to mul
     })
 })
 
-describe('qnamaker:build write dialog assets successfully if --dialog set to crosstrained', () => {
+describe('qnamaker:build write dialog and settings assets successfully if --dialog set to crosstrained', () => {
   before(async function () {
     await fs.ensureDir(path.join(__dirname, './../../../results/'))
 
@@ -310,7 +318,7 @@ describe('qnamaker:build write dialog assets successfully if --dialog set to cro
   test
     .stdout()
     .command(['qnamaker:build', '--in', './test/fixtures/testcases/qnabuild/sandwich/qnafiles/sandwich2.en-us.qna', '--subscriptionKey', uuidv1(), '--botName', 'test', '--dialog', 'crosstrained', '--out', './results', '--log', '--suffix', 'development'])
-    .it('should write dialog assets successfully when --dialog set to crosstrained', async ctx => {
+    .it('should write dialog and settings assets successfully when --dialog set to crosstrained', async ctx => {
       expect(await compareFiles('./../../../results/qnamaker.settings.development.westus.json', './../../fixtures/testcases/qnabuild/sandwich/config/qnamaker.settings.development.westus.json')).to.be.true
       expect(await compareFiles('./../../../results/sandwich2.lu.qna.dialog', './../../fixtures/testcases/qnabuild/sandwich/dialogs/sandwich2.lu.qna.dialog')).to.be.true
       expect(await compareFiles('./../../../results/sandwich2.qna.dialog', './../../fixtures/testcases/qnabuild/sandwich/dialogs/sandwich2.qna.dialog')).to.be.true
@@ -353,7 +361,7 @@ describe('qnamaker:build write crosstrained recognizer asset successfully if qna
     })
 })
 
-describe('qnamaker:build write dialog assets successfully with multi locales', () => {
+describe('qnamaker:build write dialog and settings assets successfully with multi locales', () => {
   before(async function () {
     await fs.ensureDir(path.join(__dirname, './../../../results/'))
 
@@ -379,8 +387,15 @@ describe('qnamaker:build write dialog assets successfully with multi locales', (
         qnaDocuments: [{
           id: 1,
           source: 'custom editorial',
-          questions: ['Météo'],
-          answer: 'C\'est pluvieux',
+          questions: ['who are you'],
+          answer: 'I\'m Lucy',
+          metadata: []
+        },
+        {
+          id: 2,
+          source: 'custom editorial',
+          questions: ['weather'],
+          answer: 'It\'s rainy',
           metadata: []
         }]
       })
@@ -391,15 +406,8 @@ describe('qnamaker:build write dialog assets successfully with multi locales', (
         qnaDocuments: [{
           id: 1,
           source: 'custom editorial',
-          questions: ['who are you'],
-          answer: 'I\'m Lucy',
-          metadata: []
-        },
-        {
-          id: 2,
-          source: 'custom editorial',
-          questions: ['weather'],
-          answer: 'It\'s rainy',
+          questions: ['Météo'],
+          answer: 'C\'est pluvieux',
           metadata: []
         }]
       })
@@ -419,7 +427,7 @@ describe('qnamaker:build write dialog assets successfully with multi locales', (
   test
     .stdout()
     .command(['qnamaker:build', '--in', './test/fixtures/testcases/qnabuild/multi-locales/qnafiles', '--subscriptionKey', uuidv1(), '--botName', 'test', '--dialog', 'multiLanguage', '--out', './results', '--log', '--suffix', 'development'])
-    .it('should write dialog assets successfully with multi locales', async ctx => {
+    .it('should write dialog and settings assets successfully with multi locales', async ctx => {
       expect(await compareFiles('./../../../results/qnamaker.settings.development.westus.json', './../../fixtures/testcases/qnabuild/multi-locales/config/qnamaker.settings.development.westus.json')).to.be.true
       expect(await compareFiles('./../../../results/Foo.en-us.qna.dialog', './../../fixtures/testcases/qnabuild/multi-locales/dialogs/Foo.en-us.qna.dialog')).to.be.true
       expect(await compareFiles('./../../../results/Foo.qna.dialog', './../../fixtures/testcases/qnabuild/multi-locales/dialogs/Foo.qna.dialog')).to.be.true
@@ -714,7 +722,7 @@ describe('qnamaker:build create a new knowledge base successfully with endpoint 
     })
 })
 
-describe('qnamaker:build write dialog assets successfully if schema is specified', () => {
+describe('qnamaker:build write dialog and settings assets successfully if schema is specified', () => {
   before(async function () {
     await fs.ensureDir(path.join(__dirname, './../../../results/'))
 
@@ -756,7 +764,7 @@ describe('qnamaker:build write dialog assets successfully if schema is specified
   test
     .stdout()
     .command(['qnamaker:build', '--in', './test/fixtures/testcases/qnabuild/sandwich/qnafiles/sandwich2.en-us.qna', '--subscriptionKey', uuidv1(), '--botName', 'test', '--out', './results', '--log', '--suffix', 'development', '--dialog', 'crosstrained', '--schema', 'https://raw.githubusercontent.com/microsoft/BotFramework-Composer/stable/Composer/packages/server/schemas/sdk.schema'])
-    .it('should write dialog assets successfully if schema is specified', async ctx => {      
+    .it('should write dialog and settings assets successfully if schema is specified', async ctx => {      
       expect(await compareFiles('./../../../results/qnamaker.settings.development.westus.json', './../../fixtures/testcases/qnabuild/sandwich/config/qnamaker.settings.development.westus.json')).to.be.true
       expect(await compareFiles('./../../../results/sandwich2.en-us.qna.dialog', './../../fixtures/testcases/qnabuild/sandwich/dialogs-with-schema/sandwich2.en-us.qna.dialog')).to.be.true
       expect(await compareFiles('./../../../results/sandwich2.qna.dialog', './../../fixtures/testcases/qnabuild/sandwich/dialogs-with-schema/sandwich2.qna.dialog')).to.be.true
@@ -764,7 +772,7 @@ describe('qnamaker:build write dialog assets successfully if schema is specified
     })
 })
 
-describe('qnamaker:build write dialog assets successfully when empty files exist', () => {
+describe('qnamaker:build write dialog and settings assets successfully when empty files exist', () => {
   before(async function () {
     await fs.ensureDir(path.join(__dirname, './../../../results/'))
 
@@ -826,7 +834,7 @@ describe('qnamaker:build write dialog assets successfully when empty files exist
   test
     .stdout()
     .command(['qnamaker:build', '--in', './test/fixtures/testcases/qnabuild/empty-file/qnafiles', '--subscriptionKey', uuidv1(), '--botName', 'test', '--dialog', 'crosstrained', '--out', './results', '--log', '--suffix', 'development'])
-    .it('should write dialog assets successfully when empty files exist', async ctx => {
+    .it('should write dialog and settings assets successfully when empty files exist', async ctx => {
       expect(ctx.stdout).to.contain('empty.qna loaded')
       expect(ctx.stdout).to.contain('non-empty.qna loaded')
       expect(ctx.stdout).to.contain('empty.zh-cn.qna loaded')
@@ -842,5 +850,76 @@ describe('qnamaker:build write dialog assets successfully when empty files exist
       expect(await compareFiles('./../../../results/non-empty.qna.dialog', './../../fixtures/testcases/qnabuild/empty-file/dialogs/non-empty.qna.dialog')).to.be.true
       expect(await compareFiles('./../../../results/non-empty.en-us.qna.dialog', './../../fixtures/testcases/qnabuild/empty-file/dialogs/non-empty.en-us.qna.dialog')).to.be.true
       expect(await compareFiles('./../../../results/non-empty.lu.qna.dialog', './../../fixtures/testcases/qnabuild/empty-file/dialogs/non-empty.lu.qna.dialog')).to.be.true
+    })
+})
+
+describe('qnamaker:build write settings assets only successfully if --dialog is not set', () => {
+  before(async function () {
+    await fs.ensureDir(path.join(__dirname, './../../../results/'))
+
+    nock('https://westus.api.cognitive.microsoft.com')
+      .get(uri => uri.includes('qnamaker'))
+      .reply(200, {
+        knowledgebases:
+          [{
+            name: 'test(development).en-us.qna',
+            id: 'f8c64e2a-1111-3a09-8f78-39d7adc76ec5',
+            hostName: 'https://myqnamakerbot.azurewebsites.net'
+          }]
+      })
+
+    nock('https://westus.api.cognitive.microsoft.com')
+      .get(uri => uri.includes('knowledgebases'))
+      .reply(200, {
+        qnaDocuments: [{
+          id: 1,
+          source: 'custom editorial',
+          questions: ['how many sandwich types do you have', 'how many tastes do you have'],
+          answer: '25 types',
+          metadata: []
+        }]
+      })
+
+    nock('https://westus.api.cognitive.microsoft.com')
+      .get(uri => uri.includes('endpointkeys'))
+      .reply(200, {
+        primaryEndpointKey: 'yyyyyyyyyyyyyyyyyyyyyyyyyyyyyyy',
+        secondaryEndpointKey: 'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx'
+      })
+  })
+
+  after(async function () {
+    await fs.remove(path.join(__dirname, './../../../results/'))
+  })
+
+  test
+    .stdout()
+    .command(['qnamaker:build', '--in', './test/fixtures/testcases/qnabuild/sandwich/qnafiles/sandwich2.en-us.qna', '--subscriptionKey', uuidv1(), '--botName', 'test', '--out', './results', '--log', '--suffix', 'development'])
+    .it('should write settings assets only successfully if --dialog is not set', async ctx => {      
+      expect(await compareFiles('./../../../results/qnamaker.settings.development.westus.json', './../../fixtures/testcases/qnabuild/sandwich/config/qnamaker.settings.development.westus.json')).to.be.true
+      expect(await compareFiles('./../../../results/sandwich2.en-us.qna.dialog', './../../fixtures/testcases/qnabuild/sandwich/dialogs/sandwich2.en-us.qna.dialog')).to.be.false
+      expect(await compareFiles('./../../../results/sandwich2.qna.dialog', './../../fixtures/testcases/qnabuild/sandwich/dialogs/sandwich2.qna.dialog')).to.be.false
+    })
+})
+
+describe('qnamaker:build throw qnamaker build failed exception successfully', () => {
+  before(function () {
+    nock('https://westus.api.cognitive.microsoft.com')
+      .get(uri => uri.includes('qnamaker'))
+      .reply(401, {
+        error: {
+          code: 401,
+          message: 'Access denied due to invalid subscription key.'
+        }
+      })
+  })
+
+  test
+    .stdout()
+    .stderr()
+    .command(['qnamaker:build', '--in', './test/fixtures/testcases/qnabuild/sandwich/qnafiles/sandwich.en-us.qna', '--subscriptionKey', uuidv1(), '--botName', 'test', '--log', '--suffix', 'development'])
+    .exit(1)
+    .it('should throw qnamaker build failed exception successfully', ctx => {
+      expect(ctx.stderr).to.contain('Qnamaker build failed: Access denied due to invalid subscription key.')
     })
 })
